==> code/01-read_data.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# Read data {#Read Data}
## Load packages

```{r}

source("./00_dependencies.R")

```


## Load data

```{r}

####-----  List samples
samples <- str_replace(str_sort(list.files("../data/counts_mx"), numeric = TRUE),
                       "_filtered_feature_bc_matrix",
                       "")

cat("Samples: \n", samples)


####----- Create a named vector with paths to samples
samples.dirs <- sapply(samples, function(i) {
  d10x <- file.path('../data/counts_mx',
                    paste0(i, '_filtered_feature_bc_matrix'))
  d10x
})

cat("Sample directories: \n", samples.dirs)


####----- Read count matrices
gex_data <- Read10X(data.dir = samples.dirs,
                    unique.features = TRUE,
                    strip.suffix = TRUE)
str(gex_data)

####----- Create seurat object
gex <- CreateSeuratObject(gex_data, 
                          project = "pSS",
                          min.cells = 0,
                          min.features = 0,
                          names.field = 1)

counts <- as.data.frame(table(gex@meta.data$orig.ident))
colnames(counts) <- c("sample", "cells")
write.csv(counts, "../results/GEX_cellcounts_RAW.txt")


####-----  rm tmp objects
rm(samples, samples.dirs, gex_data)

####-----  change orig.ident for P007 and P008; libraries were generated and sequenced twice
gex@meta.data$orig.ident <- gsub("P007[ab]", "P007", gex@meta.data$orig.ident)
gex@meta.data$orig.ident <- gsub("P008[ab]", "P008", gex@meta.data$orig.ident)


```


## load pheno file

```{r}


source("./00b_pheno.R")

#Four non-overlapping groups
#1. SSAB: SSA+/SSB+
#2. SSA+: SSA+/SSB-
#3. SSA-: SSA-/SSB-
#4. CTRL: CTRL


```


## add pheno info to gex

```{r}

meta.data.tmp <- gex@meta.data
meta.data.tmp$cell <- row.names(meta.data.tmp)
meta.data.tmp <- merge(meta.data.tmp, pheno, by = "orig.ident", all.x = TRUE)
row.names(meta.data.tmp) <- meta.data.tmp$cell
meta.data.tmp$cell <- NULL

gex <- AddMetaData(gex, meta.data.tmp)

Idents(object = gex) <- "orig.ident"


```


## save rds GEX1_RAW.rds

```{r include = FALSE}

saveRDS(gex, file = "../results/GEX1_RAW.rds")
#gex <- readRDS("../results/GEX1_RAW.rds")


```


## Print sessionInfo()

```{r}

sessionInfo()

```
