==> code/03-dimred1.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---


# dimensionality reduction, clusterimg, and cluster annotation {#Dimred1}
## load packages

```{r}

source("./00a_dependencies.R")

```


## load data

```{r}

gex <- readRDS("../results/GEX2_FILTERED.rds")
annot <- readRDS("../annot/refdata-gex-GRCh38-2020-A.genes.gtf.annot.rds")

```


## normalize and get variable features
 
```{r}

gex <- NormalizeData(gex, scale.factor = 10000)
gex <- FindVariableFeatures(gex, selection.method = "vst", nfeatures = 4000)

p1 <- LabelPoints(plot = VariableFeaturePlot(gex), 
                  points = head(VariableFeatures(gex), 80), 
                  repel = TRUE,
                  xnudge = 0,
                  ynudge = 0,
                  max.overlaps = Inf) + 
        theme_classic() +
        ggtitle(paste0(length(VariableFeatures(gex)), " Variable features - allCells (all features"))


#### ------ Subset VariableFeatures(gex) to only include biotype == "protein_coding"
VariableFeatures(gex) <-
  VariableFeatures(gex)[VariableFeatures(gex) %in%
                          annot$external_gene_name[annot$gene_biotype == "protein_coding"]]

p2 <- LabelPoints(plot = VariableFeaturePlot(gex), 
                  points = head(VariableFeatures(gex), 80), 
                  repel = TRUE,
                  xnudge = 0,
                  ynudge = 0,
                  max.overlaps = Inf) + 
        theme_classic() +
        ggtitle(paste0(length(VariableFeatures(gex)), " Variable features - allCells (protein_coding)"))

ggsave2("../results/GEX_variableFeatures_allCells.png", 
        plot_grid(p1, p2, ncol = 2, labels = "AUTO", align = "h"), 
        width = 30, height = 15, unit = "cm")


```


## scale data

```{r}

#### ----- Regressing out nFeature_RNA
gex <- ScaleData(gex, 
                 features = VariableFeatures(gex),
                 vars.to.regress = c("nFeature_RNA"))

```


## PCA

```{r}

#RunPCA() takes around 2h for this data set using 16 cores
gex <- RunPCA(
  gex,
  assay = "RNA",
  features = VariableFeatures(gex),
  npcs = 100,
  reduction.name = "pca_1",
  verbose = TRUE
)

#remove scale data slot to reduce Seurat object size
gex@assays$RNA@scale.data <- matrix(0)
gc()

```


## batch correction - harmony 

```{r}

#RunHarmony() takes around 1h for this data set using 16 cores

gex <- RunHarmony(
  gex,
  group.by.vars = "orig.ident",
  reduction = "pca_1",
  reduction.save = "harmony_1",
  assay = "RNA",
  project.dim = FALSE, #project.dim = FALSE needed for seurat object v4.0.0??
  verbose = TRUE
)

```


### plot PCA and harmony

```{r}

p1 <-DimPlot(object = gex,
              reduction = "pca_1",
              pt.size = .1,
              group.by = "orig.ident",
              raster = FALSE) +
                  NoLegend() +
                  ggtitle("PCA_1, by Sample")

p2 <- DimPlot(object = gex,
              reduction = "harmony_1",
              pt.size = .1,
              group.by = "orig.ident",
              raster = FALSE) +
                  NoLegend() +
                  ggtitle("Harmony_1, by Sample")

ggsave2("../results/GEX_PCA_harmony_allCells.png", 
        plot_grid(p1, p2, ncol = 2, labels = NULL, align = "h"), 
        width = 30, height = 15, unit = "cm")


```


## run UMAP

```{r}

gex <- RunUMAP(gex,
                dims = 1:50,
                reduction = "harmony_1",
                metric = "correlation",
                reduction.name = "umap_1",
                min.dist = 0.4, #local cell separation
                spread = .5, #global cell separation
                n.neighbors = 30, #30 on bianca!
                repulsion.strength = 0.4, #~2x min.dist, repulsion to cells fr annat cluster, global separation
                negative.sample.rate = 50, #ant ggr n.neighbors celler, global distance
                n.epochs = 100,
                n.components = 2
              )


```


## clustering

```{r}

gex <- FindNeighbors(gex,
  reduction = "harmony_1",
  dims = 1:50,
  k.param = 30,
  verbose = TRUE
)

gex <- FindClusters(gex, 
                    resolution = 0.4, 
                    verbose = TRUE, 
                    graph.name = "RNA_nn")


```


## cluster marker genes

```{r}

sample_size <- table(gex$seurat_clusters)
sample_size[sample_size > 100] <- 100

DGE_cells <- lapply(names(sample_size), function(x) {
  set.seed(1)
  sample(colnames(gex)[gex$seurat_clusters == x], size = sample_size[x])
})

DGE_cells <- unlist(DGE_cells)
DGE_temp<- gex[, DGE_cells]
DGE_temp <- SetIdent(DGE_temp, value = "seurat_clusters")
detable <- FindAllMarkers( DGE_temp, 
                           min.pct = 0.1, 
                           min.cells.feature = 0.1, 
                           max.cells.per.ident = 50, 
                           only.pos = TRUE, 
                           logfc.threshold = 0.1, 
                           assay = "RNA")

detable$pct.diff <- detable$pct.1 - detable$pct.2
write.csv(detable, 
          "../results/DGE_allCells_seurat_clusters.csv")

detable$log.pct.diff <- log2(detable$pct.1 / (detable$pct.2 + 0.01) )
detable <- detable[ detable$p_val < 0.05, ]

detable %>% group_by(cluster) %>% top_n(-30, p_val)  %>% top_n(20, pct.diff) -> top5
ord <-
  factor(sapply(unique(as.character(top5$gene)), function(x) {
    getcluster(DGE_temp,
               x,
               "seurat_clusters")
  }))

genes_to_plot <- unique(as.character(top5$gene))[order(as.numeric(as.character(ord)))]

pdf("../results/DGE_allCells_seurat_clusters_dotplot.pdf", width = 10, height = 50)
rafalib::mypar(1, 1, mar = c(6, 6, 1, 5))
plot_dots(gex, 
          genes_to_plot, 
          clustering = "seurat_clusters", 
          show_grid = T, 
          main = "top cluster markers", 
          cex.main = 1, font.main = 1, cex.col = 1.3, srt = 0, cex.row = 1.3)
dev.off()


```


## cluster annotation - manual

```{r}

annotation <- c(
  `B-cell` = 0,
  `B-cell` = 1,
  `B-cell` = 2,
  `B-cell` = 3,
  `B-cell` = 4,
  `B-cell` = 5,
  `B-cell` = 8,
  `T-cell` = 6,
  Monocyte = 7,
  `Plasma cell` = 9
)

gex$celltype <- names(annotation)[match(gex$seurat_clusters, annotation)]
gex$celltype <- as.factor(gex$celltype)


```


## plot
### plot SFig1A, Celltypes per sample

```{r}

p3 <- DimPlot(gex, 
               reduction = "umap_1", 
               pt.size = .01, 
               group.by = "celltype", 
               label = TRUE, 
               repel = TRUE, 
               raster = TRUE,
               cols = brewer.pal(4,"Set3")) + 
          ggtitle(paste0("")) +
          ylab("UMAP 2") +
          xlab("UMAP 1") +
          theme(axis.text = element_blank(),
                axis.ticks = element_blank(),
                legend.position = "none",
                plot.title = element_text(hjust = 0))

cells <- as_tibble(gex@meta.data) %>%
  group_by(orig.ident) %>%
  summarise(tot = length(orig.ident),
            `B-cell` = sum(celltype == "B-cell"),
            Monocyte = sum(celltype == "Monocyte"),
            `Plasma cell` = sum(celltype == "Plasma cell"),
            `T-cell` = sum(celltype == "T-cell")) 

#colSums(cells[,2:ncol(cells)])
#     tot   B-cell Monocyte   Plasma cell   T-cell 
#  242572   232166     4252      113     6041 

p4 <- cells %>% gather(celltype, cellcount, `B-cell`:`T-cell`) %>% 
 ggplot(aes(x = orig.ident, 
            y = cellcount, 
            fill = celltype)) + 
  theme_classic() + 
  geom_bar(stat = "identity", 
           color = "black", 
           position = position_stack(reverse = TRUE)) + 
  theme(legend.title = element_blank(),
        axis.text.x = element_text(face = "bold", angle = 90, vjust = 0.5, hjust = 1),
        axis.text.y = element_text(face = "bold")) +
  xlab("") +
  ylab("cell count") +
  scale_fill_manual(values = brewer.pal(4,"Set3")) +
  ggtitle(paste0("B-cells and Plasma cells: ", sum(cells$`B-cell`) + sum(cells$`Plasma cell`) , "\n", 
                 "Total cells: ", sum(cells$tot)))

ggsave2(paste0("../results/SFig1A.png"), 
        plot_grid(p3, p4, align = "hv", ncol = 2, rel_widths = c(1, 1.8)) +
        bgcolor("white") + 
        border("white")  ,
        width = 25, height = 10, unit = "cm", dpi = 300)


```


### plot SFig1B, Dimplot and markler genes for all cells


```{r}

p1 <- DimPlot(gex, 
                  reduction = "umap_1", 
                  pt.size = .05, 
                  group.by = "orig.ident", 
                  raster = TRUE) + 
          NoLegend() + 
          labs(title = paste0(nrow(gex@meta.data), " cells"),
               subtitle = "by sample") +
          ylab("UMAP 2") +
          xlab("UMAP 1") +
  theme(axis.text = element_blank(),
        axis.ticks = element_blank(),
        plot.title = element_text(hjust = 0))


p2 <- DimPlot(gex, 
                  reduction = "umap_1", 
                  pt.size = .05, 
                  group.by = "seurat_clusters", 
                  label = TRUE, 
                  repel = TRUE, 
                  raster = TRUE) + 
          NoLegend() + 
          labs(title = "",
               subtitle = "by cluster") +
          ylab("UMAP 2") +
          xlab("UMAP 1") +
  theme(axis.text = element_blank(),
        axis.ticks = element_blank(),
        plot.title = element_text(hjust = 0))


ig.features <- c("IGHA1", "IGHA2", "IGHG1", "IGHG2", "IGHG3", "IGHG4", 
                 "IGHD", "IGHE", "IGHM")

pbmc.markers <- c("MS4A1", "CD19", "CD27", "CD79A", "GNLY", 
                  "CD3E", "CD14", "LYZ", "CD8A")

pig <- list()
for (i in ig.features) {
  
  pig[[i]] <- FeaturePlot(gex, 
                        reduction = "umap_1", 
                        features = i,
                        pt.size = 0.5,
                        order = TRUE,
                        cols = c("lightgray", "blue", "navy")) + 
    xlab("") + 
    ylab("") + 
    theme(legend.position = "none",
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          plot.title = element_text(hjust = 0))
  
}

pp <- list()
for (i in pbmc.markers) {
  
  pp[[i]] <- FeaturePlot(gex, 
                        reduction = "umap_1", 
                        features = i,
                        pt.size = 0.5,
                        order = TRUE,
                        cols = c("lightgray", "blue", "navy")) + 
    xlab("") + 
    ylab("") + 
    theme(legend.position = "none",
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          plot.title = element_text(hjust = 0))
  
}

p3 <- plot_grid(p1, p2, ncol = 1)
p4 <- plot_grid(plotlist = pig, ncol = 3)
p5 <- plot_grid(plotlist = pp, ncol = 3)

ggsave2(paste0("../results/SFig1B.png"), 
        plot_grid(p3, NULL, p4, NULL, p5, ncol = 5, rel_widths = c(1, 0.1, 1.5, 0.1, 1.5)) + 
          bgcolor("white") + 
          border("white"), 
        width = 30, height = 15, unit = "cm", dpi = 300)


```


## save rds GEX3_DIMRED1.rds

```{r}

saveRDS(gex, file = "../results/GEX3_DIMRED1.rds")
#gex <- readRDS(file = "../results/GEX3_DIMRED1.rds")

```


## print sessionInfo()

```{r}

sessionInfo()

```
