==> code/04-subsetB.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# subset B-cells for further processing {#Subset_B-cells}
## soad packages

```{r}

source("./00a_dependencies.R")

```


## load data, filter to only include B cell clusters and remove sample P024 (containing a low number of B cells and plasma cells)

```{r}

gex <- readRDS("../results/GEX3_DIMRED1.rds") %>%
  filter(celltype == "B-cell" | celltype == "Plasma cell")  %>%
           filter(orig.ident != "P024")


```


## save rds GEX4_BCELLS.rds

```{r}

saveRDS(gex, file = "../results/GEX4_BCELLS.rds")
#gex <- readRDS(file = "../results/GEX4_BCELLS.rds") 


```


## print sessionInfo()

```{r}

sessionInfo()

```
