==> code/05-dimred2.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# Dimensionality reduction and clustering of B-cells {#Dimred2}
## load packages

```{r}

source("./00a_dependencies.R")

```


## load data

```{r}

gex <- readRDS("../results/GEX4_BCELLS.rds")
annot <- readRDS("../suppl/annot/refdata-gex-GRCh38-2020-A.genes.gtf.annot.rds")

```


## get variable features B-cells
 
```{r}

gex <- FindVariableFeatures(gex, 
                            selection.method = "vst", 
                            nfeatures = 4000)

p1 <- LabelPoints(plot = VariableFeaturePlot(gex), 
                  points = head(VariableFeatures(gex), 80), 
                  repel = TRUE,
                  xnudge = 0,
                  ynudge = 0,
                  max.overlaps = Inf) + 
        theme_classic() +
        ggtitle(paste0(length(VariableFeatures(gex)), " Variable features - B-cells"))


```


## subset VariableFeatures(gex) to only include biotype == "protein_coding"

```{r}

VariableFeatures(gex) <-
  VariableFeatures(gex)[VariableFeatures(gex) %in%
                          annot$external_gene_name[annot$gene_biotype == "protein_coding"]]

p2 <- LabelPoints(plot = VariableFeaturePlot(gex), 
                  points = head(VariableFeatures(gex), 80), 
                  repel = TRUE,
                  xnudge = 0,
                  ynudge = 0,
                  max.overlaps = Inf) + 
        theme_classic() +
        ggtitle(paste0(length(VariableFeatures(gex), " Variable features - B-cells (protein_coding)")))


ggsave2("../results/VariableFeatures_B-cells.png",
        plot_grid(p1, p2, ncol = 2, labels = NULL, align = "h"), 
        width = 30, height = 15, unit = "cm")


```


## scale data B-cells

```{r}

gex <- ScaleData(gex, 
                 features = VariableFeatures(object = gex),
                 vars.to.regress = c("nFeature_RNA"))

```


## PCA B-cells

```{r}

#RunPCA() takes around 2h for this data set using 16 cores

gex <- RunPCA(gex,
              assay = "RNA",
              features = VariableFeatures(object = gex),
              npcs = 100,
              reduction.name = "pca_2",
              verbose = TRUE)

gex@assays$RNA@scale.data <- matrix(0); gc() #remove scale.data to reduce object size


```


## run Harmony B-cells

```{r}

gex <- RunHarmony(
  gex,
  group.by.vars = "orig.ident",
  reduction = "pca_2",
  reduction.save = "harmony_2",
  assay = "RNA",
  project.dim = FALSE, #project.dim = FALSE needed for seurat object v4.0.0??
  verbose = TRUE
)


```


## plot PCA and harmony B-cells

```{r}

p1 <- DimPlot(object = gex, 
               reduction = "pca_2", 
               pt.size = .1, 
               group.by = "orig.ident", 
               raster = FALSE) + 
            NoLegend() + 
            ggtitle("PCA B-cells, by Sample")

p2 <- DimPlot(object = gex, 
               reduction = "harmony_2", 
               pt.size = .1, 
               group.by = "orig.ident", 
               raster=FALSE) + 
                NoLegend() + 
                ggtitle("Harmony B-cells, by Sample")

ggsave2("../results/PCA_harmony_B-cells.png", 
        plot_grid(p1, p2, ncol = 2, labels = "AUTO", align = "h"), 
        width = 30, height = 15, unit = "cm")


```


## run UMAP for B-cells

```{r}

gex <- RunUMAP(
  gex,
  dims = 1:50,
  reduction = "harmony_2",
  metric = "correlation",
  reduction.name = "umap_2",
  min.dist = 0.05, #local cell separation
  spread = .3, #global cell separation
  n.neighbors = 30,
  repulsion.strength = 0.1, #~2x min.dist, repulsion to cells fr annat cluster, global
  negative.sample.rate = 50, #ant ggr n.neighbors celler, global distance
  n.epochs = 100,
  n.components = 2
)


```


## clustering B-cells

```{r}

gex <- FindNeighbors(gex, 
                     reduction = "harmony_2", 
                     dims = 1:50,
                     k.param = 15,
                     verbose = TRUE)

gex <- FindClusters(gex, 
                    resolution = 1.5, 
                    verbose = TRUE, 
                    graph.name = "RNA_snn")

table(gex@meta.data$seurat_clusters)

pdf("../results/UMAP_B-cells_clusters.pdf", width = 12, height = 12, useDingbats = F)
mypar()
plot_meta(gex, "umap_2", "seurat_clusters", label = T)
dev.off()


```


## hclust 
### test cluster relations for B cell clusters

```{r}

cluster_means <-
  sapply(as.character(unique(gex$seurat_clusters)), function(x) {
    rowMeans(gex@assays$RNA@data[, gex$seurat_clusters == x])
  })

adj <- (1 - cor(cluster_means)) / 2
h <- hclust(as.dist(adj), method = "complete")

png("../results/hclust_clusters_B-cells.png", 
    width = 1000, 
    height = 1000, 
    units = "px",
    pointsize = 20)
par(las = 2)
plot(as.dendrogram(h))
dev.off()


```


## patient subtypes per cluster

```{r}

mypal <- RColorBrewer::brewer.pal(4,"Set3")

res <- table(list(gex$seurat_clusters, gex$patient_group))
res <- t(res)
res <- res[,order(colSums(res),decreasing = T)]
res2 <- t( t(res) / colSums(res) )
res3 <- res / rowSums(res)
res4 <- t( t(res3) / colSums(res3) )

#plot proportion of cells per cluster
png(file = paste0("../results/cluster_patient_abundance.png"),
    width = 20, height = 8, units = "cm", res = 200)
mypar(1,1,mar = c(2,4,2,5), xpd=TRUE, cex.axis = 0.7)
barplot(res2*100, las = 1, col = mypal, ylab = "Proportion of cells\n")
legend("topright", legend = rownames(res2), bty = "n",fill = mypal, inset=c(-0.15,0))
dev.off()

#plot proportion of cells per cluster
png(file = paste0("../results/cluster_patient_cellcounts_abundance.png"),
    width = 20, height = 8, units = "cm", res = 200)
mypar(2,1,mar = c(2,4,2,5), xpd=TRUE, cex.axis = 0.7)
barplot(res, las = 1, col = mypal, ylab = "Number of cells\n")
barplot(res2*100, las = 1, col = mypal, ylab = "Proportion of cells\n")
legend("topright", legend = rownames(res), bty = "n",fill = mypal, inset=c(-0.15,0))
dev.off()

p1 <- DimPlot(gex, split.by = "patient_group", reduction = "umap_2", ncol = 2, raster = FALSE)
ggsave2("../results/UMAP_B-cells_patientGroup.png", 
        p1, 
        width = 30, height = 20, unit = "cm")


```


## save rds GEX5_DIMRED2.rds

```{r}

saveRDS(gex, file = "../results/GEX5_DIMRED2.rds")
#gex <- readRDS(file = "../results/GEX5_DIMRED2.rds")

```


## print sessionInfo()

```{r}

sessionInfo()

```
