==> code/06-AnnotateB-cells.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---


# annotation of B-cell clusters {#B-cell_annotation}
## load packages

```{r}

source("./00a_dependencies.R")

```


## load data

```{r}

gex <- readRDS("../results/GEX5_DIMRED2.rds")

```

## colors - cellType 

```{r}

source("./00c_colors.R")

```


## get B-cell cluster marker genes

```{r}

sample_size <- table(gex$seurat_clusters)
sample_size[sample_size > 100] <- 100
DGE_cells <- lapply(names(sample_size), function(x) {
  set.seed(1)
  sample(colnames(gex)[gex$seurat_clusters == x], size = sample_size[x])
})
DGE_cells <- unlist(DGE_cells)
DGE_temp<- gex[, DGE_cells]
DGE_temp <- SetIdent(DGE_temp, value = "seurat_clusters")
detable <- FindAllMarkers( DGE_temp, 
                           min.pct = 0.1, 
                           min.cells.feature = 0.1, 
                           max.cells.per.ident = 200, 
                           only.pos = TRUE, 
                           logfc.threshold = 0.1, 
                           assay = "RNA")

detable$pct.diff <- detable$pct.1 - detable$pct.2
write.csv(detable, 
          paste0("../results/DGE_B-cells_seurat_clusters.csv"))
detable <- read.csv("../results/DGE_B-cells_seurat_clusters.csv", stringsAsFactors = FALSE, sep = ";")

detable <- detable[detable$p_val < 0.05,  ]
detable$log.pct.diff <- log2(detable$pct.1 / (detable$pct.2 + 0.01) )

detable %>% group_by(cluster) %>%  top_n(-30, p_val)  %>% top_n(20, log.pct.diff) -> top5
ord <- factor(sapply(unique(as.character(top5$gene)),
                     function(x) {
                       getcluster(DGE_temp, x, "seurat_clusters")
                     }))

genes_to_plot <- unique(as.character(top5$gene))[order(as.numeric( as.character(ord) ))]

pdf(paste0("../results/DGE_B-cells_seurat_clusters_dotplot.pdf"), 
    width = 10, height = length(genes_to_plot) / 4 + 3 )
rafalib::mypar(1, 1, mar = c(6, 6, 1, 5))
plot_dots(DGE_temp, 
          genes_to_plot, 
          clustering = "seurat_clusters", 
          show_grid = TRUE, 
          main = "top cluster markers", 
          cex.main = 1, font.main = 1, cex.col = 1.3, srt = 0, cex.row = 1.3)
dev.off()


```



## integrate with stewart et al. 2021 per cell using scPred()

```{r}
#https://pubmed.ncbi.nlm.nih.gov/33815362/

#Transitional (CD19+IgD+CD27-CD10+)
#Naive (CD19+IgD+CD27-CD10-)
#IgM Memory (CD19+IgD+CD27+)
#Classical Memory (CD19+IgD-CD27+)
#Double Negative (CD19+IgD-CD27-)

stewart <- readRDS("../suppl/Stewart_et_al_2021/E-MTAB-9544_rds/scPure2_HB6_UMAP3D.rds")

#filter and process
stewart <- subset(stewart, subset = nFeature_RNA > 200 & nFeature_RNA < 3000 & percent.mt < 10)
stewart <- stewart[rowSums(stewart@assays$RNA@counts) != 0, ]

stewart <- stewart %>%
   NormalizeData() %>%
   FindVariableFeatures() %>%
   ScaleData() %>%
   RunPCA() %>%
   RunUMAP(dims = 1:50)


####------------- per sorted cell type

#subset Stewart seurat object
Idents(object = stewart) <- "library_id"
stewart.lid.300 <- subset(stewart, downsample = 300)

#train model
ref1 <- getFeatureSpace(stewart.lid.300, "library_id")
ref1 <- trainModel(ref1)
get_scpred(ref1)

# |Cell type        |   n| Features|Method    |   ROC|  Sens|  Spec|
# |:----------------|---:|--------:|:---------|-----:|-----:|-----:|
# |Classical_Memory | 300|       50|svmRadial | 0.967| 0.803| 0.974|
# |DN               | 300|       50|svmRadial | 0.933| 0.677| 0.960|
# |IgM_Memory       | 300|       50|svmRadial | 0.963| 0.790| 0.965|
# |Naive            | 300|       50|svmRadial | 0.947| 0.720| 0.960|
# |Transitional     | 300|       50|svmRadial | 0.984| 0.873| 0.980|

saveRDS(ref1, "../results/stewart_ref1_library_id.rds")
#ref1 <- readRDS("../results/stewart_ref1_cluster.rds")

#predict cell types
gex <- scPredict(gex, ref1)

#rename columns in seurat meta.data 
scPredcols <- c("scpred_max", "scpred_prediction", "scpred_no_rejection")
names(gex@meta.data)[names(gex@meta.data) == scPredcols[1]] <- paste0(scPredcols[1], "_ref1")
names(gex@meta.data)[names(gex@meta.data) == scPredcols[2]] <- paste0(scPredcols[2], "_ref1")
names(gex@meta.data)[names(gex@meta.data) == scPredcols[3]] <- paste0(scPredcols[3], "_ref1")
names(gex@meta.data)[names(gex@meta.data) == "scpred_Naive"] <- "scpred_Naive_"

rm(ref1); rm(stewart.lid.300)

####------------- per stewart cluster
#subset seurat object
Idents(object = stewart) <- "cluster"
stewart.cl.300 <- subset(stewart, downsample = 300)
#rm(stewart)

#train model
ref2 <- getFeatureSpace(stewart.cl.300, "cluster")
ref2 <- trainModel(ref2)
get_scpred(ref2)

# |Cell type |   n| Features|Method    |   ROC|  Sens|  Spec|
# |:---------|---:|--------:|:---------|-----:|-----:|-----:|
# |Trans     | 300|       50|svmRadial | 0.999| 0.950| 0.993|
# |Naive     | 300|       50|svmRadial | 0.984| 0.817| 0.986|
# |M-mem1    | 300|       50|svmRadial | 0.978| 0.780| 0.978|
# |M-mem2    | 300|       50|svmRadial | 0.944| 0.527| 0.970|
# |C-mem1    | 300|       50|svmRadial | 0.959| 0.617| 0.977|
# |C-mem2    | 300|       50|svmRadial | 0.985| 0.763| 0.982|
# |DN1       | 300|       50|svmRadial | 0.972| 0.733| 0.987|
# |DN2       | 300|       50|svmRadial | 0.998| 0.930| 0.993|
# |DN3       | 300|       50|svmRadial | 0.914| 0.437| 0.974|
# |DN4       | 300|       50|svmRadial | 0.969| 0.690| 0.984|

saveRDS(ref2, "../results/stewart_ref2_cluster.rds")
#ref2 <- readRDS("../results/stewart_ref2_cluster.rds")
#rm(stewart.cl.300)

#predict cell types
gex <- scPredict(gex, ref2)

#rename columns in seurat meta.data 
names(gex@meta.data)[names(gex@meta.data) == scPredcols[1]] <- paste0(scPredcols[1], "_ref2")
names(gex@meta.data)[names(gex@meta.data) == scPredcols[2]] <- paste0(scPredcols[2], "_ref2")
names(gex@meta.data)[names(gex@meta.data) == scPredcols[3]] <- paste0(scPredcols[3], "_ref2")


```


## add cellType predictions to seurat object

```{r}

####---- add cluster cell type classification based on majority fraction of predicted cell type in cluster
annot.df <- data.frame(cluster = rownames(hm1), 
                       annot_sort = colnames(hm1)[apply(hm1, 1, which.max)], 
                       annot_clust = colnames(hm2)[apply(hm2, 1, which.max)])
  
gex@meta.data["cluster_stewart_sort"] <- annot.df$annot_sort[match(gex$seurat_clusters,
                                                                   annot.df$cluster)]

gex@meta.data["cluster_stewart_cl"] <- annot.df$annot_clust[match(gex$seurat_clusters,
                                                                  annot.df$cluster)]


```


## manual annotation using output from scPred and cluster marker genes

```{r}

annotation_fine <- c(
  Naive = 0,#CCR7, BACH2, SELL, IL4R
  Naive = 1,#IGHD,FCER2
  Naive = 2,#IGHD,FCER2
  Memory_IgM = 3,
  Naive_Transitional = 4,# CD9, IGHD,FCER2
  Memory_Classical = 5,
  Naive_IFN = 6,#CCR7, BACH2, SELL, IL4R, IGHD, FCER2, + IFN regulated genes
  Memory_DN2 = 7, #IGHA1
  Memory_Str_Cl = 8,
  Memory_DN1_ID3 = 9, #ID3, LGALS3, ITGB1, IGHA1
  Naive = 10,#COL19A1, CCR7, BACH2, SELL, IL4R, IGHD,FCER2
  Memory_DN4 = 11, #IGHE, IL13RA (allergy)
  Memory_IgM = 12, #?? early memory?, no concrete marker, down regulates "naive genes"
  Memory_DN2_CXCR3 = 13, #ITGB7, IGHG1, IGHG2, IGHG3, CXCR3, IGHA1
  Memory_IgM_CD1C = 14,#CD1C
  Memory_DN2_ITGAX = 15, #FCRL3, MPP6, ARL4D, ZEB2, ITGAX, ITGB7, IGHA1
  Memory_Classical = 16, #IGHA
  Memory_IgM_ALOX5 = 17, #?? ALOX5, ACADM, GID8, CMSS1
  Memory_Str_Cl_CD69 = 18, #FOS, CD69, EGR1, DUSP1
  Memory_Platelet = 19, #CCL5, F13A1, PPBP, NRGN, CLU
  Naive = 20,#
  Memory_Str_Cl = 21,#dead/dying cells
  Plasma_cell = 22
)

gex@meta.data["cluster_cellType_manual_fine"] <- names(annotation_fine)[match(gex$seurat_clusters,
                                                                     annotation_fine)] 

annotation_main <- c(
  Naive = 0,#CCR7, BACH2, SELL, IL4R
  Naive = 1,#IGHD,FCER2
  Naive = 2,#IGHD,FCER2
  Memory_IgM = 3,
  Transitional = 4,# CD9, IGHD,FCER2
  Memory_Classical = 5,
  Naive = 6,#CCR7, BACH2, SELL, IL4R, IGHD, FCER2, + IFN regulated genes
  Memory_DN = 7, #IGHA1
  Memory_Stressed = 8,
  Memory_DN = 9, #ID3, LGALS3, ITGB1, IGHA1
  Naive = 10,#COL19A1, CCR7, BACH2, SELL, IL4R, IGHD,FCER2
  Memory_DN = 11, #IGHE, IL13RA (allergy)
  Memory_IgM = 12, #?? early memory?, no concrete marker, down regulates "naive genes"
  Memory_DN = 13, #ITGB7, IGHG1, IGHG2, IGHG3, CXCR3, IGHA1
  Memory_IgM = 14,#CD1C
  Memory_DN = 15, #FCRL3, MPP6, ARL4D, ZEB2, ITGAX, ITGB7, IGHA1
  Memory_Classical = 16, #IGHA
  Memory_IgM = 17, #?? ALOX5, ACADM, GID8, CMSS1
  Memory_Stressed = 18, #FOS, CD69, EGR1, DUSP1
  Memory_Platelet = 19, #CCL5, F13A1, PPBP, NRGN, CLU
  Naive = 20,#
  Memory_Stressed = 21,#dead/dying cells
  Plasma_cell = 22
)

gex@meta.data["cluster_cellType_manual_main"] <- names(annotation_main)[match(gex$seurat_clusters,
                                                                     annotation_main)] 


```

## plots
### plot SFig2, reference integration

```{r}

p1 <- DimPlot(gex, 
              group.by = "scpred_prediction_ref1", 
              reduction = "umap_2",
              cols = unname(colors_scpred1)) + 
        xlab("UMAP 1") + 
        ylab("UMAP 2") +
        labs(title = "predicted cell type", 
             subtitle = "by sorted cell type (Stewart et al. 2021)") +
        theme(axis.text = element_blank(),
              axis.ticks = element_blank(),
              plot.title = element_text(hjust = 0.5),
              plot.subtitle = element_text(hjust = 0.5))

p2 <- DimPlot(gex, 
              group.by = "scpred_prediction_ref2", 
              reduction = "umap_2",
              cols = unname(colors_scpred2)) + 
        xlab("UMAP 1") + 
        ylab("UMAP 2") +
        labs(title = "predicted cell type", 
             subtitle = "by sorted cell type (Stewart et al. 2021)") +
        theme(axis.text = element_blank(),
              axis.ticks = element_blank(),
              plot.title = element_text(hjust = 0.5),
              plot.subtitle = element_text(hjust = 0.5))


####---- check cluster overlap with cell types
#per sorted cell type
hm1 <- round(sweep(as.data.frame.matrix(table(gex$seurat_clusters, 
                                              gex$scpred_prediction_ref1)), 
                   1, table(gex$seurat_clusters), `/`) * 100, 
             digits = 1)

phm1 <- pheatmap(hm1, 
         cluster_rows = FALSE,
         cluster_cols = FALSE,
         main = "sorted cell type, % of seurat cluster", 
         display_numbers = TRUE,
         filename = "../results/cellType_B-cells_Stewart_et_al_1_1.png", 
         number_format = "%.1f",
         legend = FALSE)

#per sorted cell type cluster
hm2 <- round(sweep(as.data.frame.matrix(table(gex$seurat_clusters, 
                                              gex$scpred_prediction_ref2)), 
                   1, table(gex$seurat_clusters), `/`) * 100, 
             digits = 1)

phm2 <- pheatmap(hm2, 
         cluster_rows = FALSE, 
         cluster_cols = FALSE,
         main = "clustered sorted cell type, % of seurat cluster", 
         display_numbers = TRUE,
         filename = "../results/cellType_B-cells_Stewart_et_al_2_2.png",
         number_format = "%.1f", 
         legend = FALSE) 

p3 <- tableGrob(table(gex$scpred_prediction_ref1, gex$scpred_prediction_ref2))

p4 <- plot_grid(p1, p2, phm1[[4]], phm2[[4]], ncol = 2, labels = NULL, align = "hv") + 
          bgcolor("white") + 
          border("white")

ggsave2(paste0("../results/SFig2.png"), 
        plot_grid(p4, p3, ncol = 1, labels = NULL, rel_heights = c(4, 1)) + 
          bgcolor("white") + 
          border("white"), 
        width = 30, height = 30, unit = "cm")

ggsave2(paste0("../results/SFig2.pdf"), 
        plot_grid(p4, p3, ncol = 1, labels = NULL, rel_heights = c(4, 1)) + 
          bgcolor("white") + 
          border("white"), 
        width = 30, height = 30, unit = "cm")


```


### plot Fig1B-C, fine cell types and IGH constant region expression (cellTypeFine)

```{r}

p1 <- DimPlot(gex, 
              reduction = "umap_2", 
              group.by = "cluster_cellType_manual_fine",
              pt.size = 0.1, 
              cols = unname(colors_cellTypeFine),
              shuffle = TRUE,
              raster = TRUE,
              raster.dpi = c(1024, 1024)) + 
  xlab("UMAP 1") + 
  ylab("UMAP 2") + 
    theme(plot.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.position = "bottom",
          legend.text = element_text(size = 6.5, face = "bold" ))

ig.features <- c("IGHA1", "IGHA2", "IGHG1", "IGHG2", "IGHG3", "IGHG4",  "IGHD", "IGHE", "IGHM")

p <- list()
for (i in ig.features) {
  
  p[[i]] <- FeaturePlot(gex, 
                        reduction = "umap_2", 
                        features = i,
                        pt.size = 0.2,
                        order = TRUE,
                        cols = c("lightgray", "blue", "navy")) + 
    theme_void() +
    theme(legend.position = "none",
          plot.title = element_text(hjust = 0.5, vjust = 0),
          plot.margin = unit(c(0, 0, 0, 0), "cm"))
  
}

p2 <- plot_grid(plotlist = p, ncol = 3, align = "hv", scale = 1.1) 



ggsave2(paste0("../results/Fig1B-C.pdf"), 
        plot_grid(p1, p2, 
                  ncol = 2, 
                  rel_widths = c(1.1, 1)) + 
          bgcolor("white") + 
          border("white"), 
        width = 29, height = 15, unit = "cm", dpi = 300)

p1 <- DimPlot(gex, 
              reduction = "umap_2", 
              group.by = "cluster_cellType_manual_fine",
              pt.size = 0.05,
              shuffle = TRUE,
              cols = unname(colors_cellTypeFine),
              raster = FALSE) + 
  xlab("UMAP 1") + 
  ylab("UMAP 2") + 
    theme(plot.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.position = "bottom",
          legend.text = element_text(size = 6.5, face = "bold" ))

ggsave2(paste0("../results/Fig1B-C.png"), 
        plot_grid(p1, p2, 
                  ncol = 2, 
                  rel_widths = c(1.1, 1)) + 
          bgcolor("white") + 
          border("white"), 
        width = 43.5, height = 22.5, unit = "cm", dpi = 300)


```


### plot Fig1D, dotplot for cluster marker genes (cellTypeFine)

```{r}

cl_genes <- c("CCR7", "BACH2", "SELL", "IL4R", "IGHD", "FCER2", "CD9", "SELL", "IGHA1", "ID3", "LGALS3", "ITGB1", "COL19A1", "IGHE",  "ITGB7", "IGHG1", "IGHG2", "IGHG3", "CXCR3", "CD1C", "FCRL3", "MPP6", "ARL4D", "ZEB2", "ITGAX", "ALOX5", "ACADM", "GID8", "CMSS1", "FOS", "CD69", "EGR1", "DUSP1", "CCL5", "F13A1", "PPBP", "NRGN", "CLU")

cl_genes <- unique(cl_genes)

p <- DotPlot(gex, 
        features = cl_genes, 
        group.by = "cluster_cellType_manual_fine",
        dot.scale = 3.5,
        cols = c("lightgray", "blue", "navy")
        ) +
  theme_classic() +
  xlab("") + 
  ylab("") + 
  coord_flip() + 
  theme(axis.text.x = element_text(angle = 90, 
                                       hjust = 1, 
                                       vjust = 0.5,
                                       size = 8),
        axis.text.y = element_text(size = 7),
        legend.text = element_text(size = 8),
        legend.title = element_text(size = 9),
        panel.grid.major = element_line(colour="grey", size = rel(0.5)))

ggsave2(paste0("../results/Fig1D.png"), 
        p, 
        width = 11, height = 11.5, unit = "cm") 

ggsave2(paste0("../results/Fig1D.pdf"), 
        p, 
        width = 11, height = 11.5, unit = "cm") 


```


###  plot Fig1E, patient subtypes per cell type main

```{r}

p1 <- ggplot(gex@meta.data[gex@meta.data$cluster_cellType_manual_main %in% c("Memory_Classical",
                                                                "Memory_DN",
                                                                "Memory_IgM",
                                                                "Naive",
                                                                "Transitional"), ], 
             aes(x = reorder(cluster_cellType_manual_main,
                             -table(cluster_cellType_manual_main)[cluster_cellType_manual_main]),
                 fill = patient_group)) +
        theme_classic() +
        geom_bar(position = position_stack(reverse = FALSE)) + 
        ylab("cells") + 
        xlab("") + 
        theme(legend.position = "none",
              axis.text.x = element_blank(),
              axis.ticks.x=element_blank()) +
        scale_fill_manual(name = "Patient group", values = scales::hue_pal()(4))

p2 <- ggplot(gex@meta.data[gex@meta.data$cluster_cellType_manual_main %in% c("Memory_Classical",
                                                                "Memory_DN",
                                                                "Memory_IgM",
                                                                "Naive",
                                                                "Transitional"), ], 
             aes(x = reorder(cluster_cellType_manual_main,
                             -table(cluster_cellType_manual_main)[cluster_cellType_manual_main]), 
                 fill = patient_group)) +
        theme_classic() +
        geom_bar(position = "fill") + 
        ylab("proportion of cells") + 
        xlab("") + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
              axis.ticks.x = element_blank(),
              legend.position = "right") +
        scale_fill_manual(name = "group", values = scales::hue_pal()(4))

#get legend
p_legend_r <- get_legend(p2) %>% ggpubr::as_ggplot()
p2 <- p2 + theme(legend.position = "bottom")
p_legend_b <- get_legend(p2) %>% ggpubr::as_ggplot()
p2 <- p2 + theme(legend.position = "none")

p3 <- plot_grid(p1, p2, 
                ncol = 1, labels = NULL, rel_heights = c(0.6, 1) , align = "v") + 
          bgcolor("white") + 
          border("white")

ggsave2("../results/Fig1E.png", 
        plot_grid(p3, p_legend_r, 
                  ncol = 2, 
                  rel_widths = c(5, 1), 
                  labels = NULL, 
                  align = "h") + 
          bgcolor("white") + 
          border("white"), 
        width = 11, height = 8, unit = "cm")

ggsave2("../results/Fig1E.pdf", 
        plot_grid(p3, p_legend_r, 
                  ncol = 2, 
                  rel_widths = c(5, 1), 
                  labels = NULL, 
                  align = "h") + 
          bgcolor("white") + 
          border("white"), 
        width = 11, height = 8, unit = "cm")


```


### plot Fig1F, violins split by cell type main

```{r}

# violin plot per celltype main per orig.ident/patient_group
cell_summary <- as_tibble(gex@meta.data) %>%
  select(orig.ident, cluster_cellType_manual_main, patient_group) %>%
  add_count(orig.ident, cluster_cellType_manual_main, .drop = FALSE) %>%
  distinct() %>%
  group_by(orig.ident) %>%
  mutate(per =  prop.table(n)) %>%
  drop_na() %>% as_data_frame()

cell_summary <- cell_summary[cell_summary$cluster_cellType_manual_main %in% c("Memory_Classical", 
                                                                              "Memory_DN",
                                                                              "Memory_IgM",
                                                                              "Naive",
                                                                              "Transitional"), ]


# reorder factor for celltypeFine to match the Fig1E
theorder <- unique(names(sort(table(gex@meta.data$cluster_cellType_manual_main)[gex@meta.data$cluster_cellType_manual_main], decreasing = TRUE)))
theorder <- theorder[theorder %in% c("Memory_Classical", 
                                     "Memory_DN",
                                     "Memory_IgM",
                                     "Naive",
                                     "Transitional")]
cell_summary$cluster_cellType_manual_main <- factor(cell_summary$cluster_cellType_manual_main,
                                         levels = theorder)

p_out <- list()
for (i in levels(cell_summary$cluster_cellType_manual_main)){
maxval  <- max(cell_summary$per[cell_summary$cluster_cellType_manual_main == i])
p_out[[i]] <- ggplot(cell_summary[cell_summary$cluster_cellType_manual_main == i, ], 
                     aes(x = patient_group,
                         y = per,
                         fill = patient_group)) +
                     theme_classic() +
                     geom_violin(scale = "width", 
                                 draw_quantiles = c(0.25, 0.5, 0.75),
                                 trim = TRUE,
                                 alpha = 0.8) +
                     geom_jitter(size = 0.3) +
                     ylab("") + 
                     xlab("") +
                     ggtitle(i) + 
                     scale_fill_manual(values = scales::hue_pal()(4)) +
                     stat_compare_means(method = "t.test", 
                                        comparisons = list(c("CTRL", "SSA-"), 
                                                           c("CTRL", "SSA+"), 
                                                           c("CTRL", "SSAB")), 
                                        label = "p",
                                        size = 3.5 #,
                                        #label.y = c(29, 35, 40)
                                        #vjust = 0.5
                                        ) +
                     theme(axis.text.x = element_text(angle = 90, 
                                                      vjust = 0.5, 
                                                      hjust = 1),
                           plot.title = element_text(size = 9,
                                                     face = "bold"),
                     legend.position = "none") +
                     scale_y_continuous(labels = scales::percent_format(accuracy = 1L), 
                                        limits = c(0, (maxval * 1.4)))
}

#draw y label
y_label <- ggdraw() + 
  draw_label("proportion of cells in sample", 
             hjust = 0.5, 
             angle = 90, 
             vjust = 0)

p4 <- plot_grid(plotlist = p_out, 
                ncol = 5, 
                labels = NULL, 
                align = "hv") + 
          bgcolor("white") + 
          border("white")

ggsave2("../results/Fig1F.png", 
        plot_grid(y_label, p4, 
                  ncol = 2, 
                  labels = NULL, 
                  align = "hv", 
                  rel_widths = c(0.03, 1)) + 
          bgcolor("white") + 
          border("white"), 
        width = 25, height = 9, unit = "cm")

ggsave2("../results/Fig1F.pdf", 
        plot_grid(y_label, p4, 
                  ncol = 2, 
                  labels = NULL, 
                  align = "hv", 
                  rel_widths = c(0.03, 1)) + 
          bgcolor("white") + 
          border("white"), 
        width = 25, height = 9, unit = "cm")


```


### plot Fig1E-F

```{r}

p5 <- plot_grid(p3, NULL, y_label, p4, 
                  ncol = 4, 
                  labels = NULL, 
                  align = "hv", 
                  rel_widths = c(0.4, 0.08, 0.05, 1.5)) + 
          bgcolor("white") + 
          border("white")


ggsave2("../results/Fig1E-F.png", 
        plot_grid(p5, p_legend_b, 
                  ncol = 1, 
                  labels = NULL, 
                  align = "hv", 
                  rel_heights = c(1, 0.1)) + 
          bgcolor("white") + 
          border("white"), 
        width = 30, height = 9, unit = "cm")

ggsave2("../results/Fig1E-F.pdf", 
        plot_grid(p5, p_legend_b, 
                  ncol = 1, 
                  labels = NULL, 
                  align = "hv", 
                  rel_heights = c(1, 0.1)) + 
          bgcolor("white") + 
          border("white"), 
        width = 30, height = 9, unit = "cm")


```


### plot SFig3A, cellTypesMain by patient group

```{r}

p <- DimPlot(gex, 
             reduction = "umap_2", 
             group.by = "cluster_cellType_manual_main",
             split.by = "patient_group", 
             cols = unname(colors_cellTypeMain),
             ncol = 2) + 
  xlab("UMAP 1") + 
  ylab("UMAP 2") + 
  #ggtitle("celltypes main") + 
    theme(axis.text = element_blank(),
          axis.ticks = element_blank(),
          plot.title = element_blank())

ggsave2(paste0("../results/SFig3A.png"), 
        p, 
        width = 20, height = 15, unit = "cm")

ggsave2(paste0("../results/SFig3A.pdf"), 
        p, 
        width = 20, height = 15, unit = "cm")


```


### plot SFig3B, KNN clusters

```{r}

gex$cluster_cluster_cellType_manual_fine <- paste(gex$seurat_clusters, 
                                                  gex$cluster_cellType_manual_fine, 
                                                  sep = "_")

gex$cluster_cluster_cellType_manual_fine <- factor(gex$cluster_cluster_cellType_manual_fine, 
                                                   levels = str_sort(unique(gex$cluster_cluster_cellType_manual_fine), numeric = TRUE))
names(table(gex$cluster_cluster_cellType_manual_fine))


p <- DimPlot(gex, 
             reduction = "umap_2", 
             group.by = "cluster_cluster_cellType_manual_fine", 
             pt.size = 0.005,
             label = TRUE, 
             repel = TRUE, 
             label.box = TRUE, 
             label.size = 3,
             cols = unname(colors_clusters),
             raster = FALSE) + 
  xlab("UMAP 1") + 
  ylab("UMAP 2") + 
  ggtitle("") + 
    theme(axis.text = element_blank(),
          axis.ticks = element_blank(),
          legend.position = "none")


ggsave2(paste0("../results/SFig3B.png"), 
        p, 
        width = 18, height = 17, unit = "cm")

ggsave2(paste0("../results/SFig3B.pdf"), 
        p, 
        width = 18, height = 17, unit = "cm")


gex$cluster_cluster_cellType_manual_fine <- NULL


```


### plot SFig3C, celltype fine per patient subgroup 

```{r}

p <- ggplot(gex@meta.data, 
            aes(x = patient_group, 
                fill =  cluster_cellType_manual_fine)) +
        scale_fill_manual(values = unname(colors_cellTypeFine)) +
        theme_classic() +
        geom_bar(position = "fill") + 
        ylab("proportion of cells") + 
        xlab("") +
        ggtitle("celltype fine per patient subgroup") +
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
              legend.title = element_blank())

ggsave2("../results/SFig3C.png", 
        p, 
        width = 20, height = 14, unit = "cm")

ggsave2("../results/SFig3C.pdf", 
        p, 
        width = 20, height = 14, unit = "cm")


```


###  plot SFig3D, patient subtypes per cell type fine

```{r}

p1 <- ggplot(gex@meta.data, aes(x = reorder(cluster_cellType_manual_fine,
                                        -table(cluster_cellType_manual_fine)[cluster_cellType_manual_fine]),
                                            fill = patient_group)) +
        theme_classic() +
        geom_bar(position = position_stack(reverse = FALSE)) + 
        ylab("cells") + 
        xlab("") + 
        theme(legend.position = "none",
              axis.text.x = element_blank(),
              axis.ticks.x=element_blank()) +
        scale_fill_manual(name = "Patient group", values = scales::hue_pal()(4))

p2 <- ggplot(gex@meta.data, aes(x = reorder(cluster_cellType_manual_fine,
                                        -table(cluster_cellType_manual_fine)[cluster_cellType_manual_fine]), 
                                            fill = patient_group)) +
        theme_classic() +
        geom_bar(position = "fill") + 
        ylab("proportion of cells") + 
        xlab("") + 
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
              axis.ticks.x = element_blank(),
              legend.position = "bottom") +
        scale_fill_manual(name = "", values = scales::hue_pal()(4))

#get legend
p_legend <- get_legend(p2) %>% ggpubr::as_ggplot()
p2 <- p2 + theme(legend.position = " none")

p3 <- plot_grid(p1, p2, 
                ncol = 1, labels = NULL, rel_heights = c(0.6, 1) , align = "v") + 
          bgcolor("white") + 
          border("white")

ggsave2("../results/SFig3D.png", 
        plot_grid(p3, p_legend, 
                  ncol = 1, 
                  rel_heights = c(1, 0.1), 
                  labels = NULL, 
                  align = "h") + 
          bgcolor("white") + 
          border("white"), 
        width = 25, height = 10, unit = "cm")

ggsave2("../results/SFig3D.pdf", 
        plot_grid(p3, p_legend, 
                  ncol = 1, 
                  rel_heights = c(1, 0.1), 
                  labels = NULL, 
                  align = "h") + 
          bgcolor("white") + 
          border("white"), 
        width = 25, height = 10, unit = "cm")


```


### plot SFig3E, violins split by cell type fine

```{r}

# violin plot per celltype fine per orig.ident/patient_group
cell_summary <- as_tibble(gex@meta.data) %>%
  select(orig.ident, cluster_cellType_manual_fine, patient_group) %>%
  add_count(orig.ident, cluster_cellType_manual_fine, .drop = FALSE) %>%
  distinct() %>%
  group_by(orig.ident) %>%
  mutate(per =  prop.table(n)) %>%
  drop_na()

# reorder factor for celltypeFine to match the SFig3D
theorder <- unique(names(sort(table(gex@meta.data$cluster_cellType_manual_fine)[gex@meta.data$cluster_cellType_manual_fine], decreasing = TRUE)))
cell_summary$cluster_cellType_manual_fine <- factor(cell_summary$cluster_cellType_manual_fine,
                                         levels = theorder)

p_out <- list()
for (i in levels(cell_summary$cluster_cellType_manual_fine)){
maxval  <- max(cell_summary$per[cell_summary$cluster_cellType_manual_fine == i])
p_out[[i]] <- ggplot(cell_summary[cell_summary$cluster_cellType_manual_fine == i, ], 
                     aes(x = patient_group,
                         y = per,
                         fill = patient_group)) +
                     theme_classic() +
                     geom_violin(scale = "width", 
                                 draw_quantiles = c(0.25, 0.5, 0.75),
                                 trim = TRUE,
                                 alpha = 0.8) + 
                     geom_jitter(size = 0.3) +
                     ylab("") + 
                     xlab("") +
                     ggtitle(i) + 
                     scale_fill_manual(values = scales::hue_pal()(4)) +
                     stat_compare_means(method = "t.test", 
                                        comparisons = list(c("CTRL", "SSA-"), 
                                                           c("CTRL", "SSA+"), 
                                                           c("CTRL", "SSAB")), 
                                        label = "p.format",
                                        size = 3.5 
                                        #vjust = 0.5
                                        ) +
                     theme(axis.text.x = element_text(angle = 90, 
                                                      vjust = 0.5, 
                                                      hjust = 1,
                                                      face = "bold"),
                           plot.title = element_text(size = 9,
                                                     face = "bold"),
                     legend.position = "none") +
                     scale_y_continuous(labels = scales::percent_format(accuracy = 1L), 
                                        limits = c(0, (maxval * 1.4)))
}

#draw y label
y_label <- ggdraw() + 
  draw_label("proportion of cells in sample", 
             fontface = 'bold',
             hjust = 0.5, 
             angle = 90, 
             vjust = 0)

#plot grid
p1 <- plot_grid(plotlist = p_out, 
                ncol = 8, 
                labels = NULL, 
                align = "hv") + 
          bgcolor("white") + 
          border("white")

#add y label
p2 <- plot_grid(y_label, p1, 
                  ncol = 2, 
                  labels = NULL, 
                  align = "hv", 
                  rel_widths = c(0.03, 1)) + 
          bgcolor("white") + 
          border("white")

ggsave2("../results/SFig3E.png", 
        plot_grid(p2, p_legend, 
                ncol = 1, 
                labels = NULL, 
                align = "hv",
                rel_heights = c(1, 0.1)) + 
          bgcolor("white") + 
          border("white"), 
        width = 40, height = 15, unit = "cm")

ggsave2("../results/SFig3E.pdf", 
         plot_grid(p2, p_legend, 
                ncol = 1, 
                labels = NULL, 
                align = "hv",
                rel_heights = c(1, 0.1)) + 
          bgcolor("white") + 
          border("white"), 
        width = 40, height = 15, unit = "cm")


```


## save rds GEX6_BANNOT.rds

```{r}

saveRDS(gex, file = "../results/GEX6_BANNOT.rds")
#gex <- readRDS(file = "../results/GEX6_BANNOT.rds")

```


## print sessionInfo()

```{r}

sessionInfo()

```
