==> code/08-DEtests.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# DE tests and Visualization {#DE-tests}
## load packages

```{r}

source("./00a_dependencies.R")

```


## load data

```{r}

gex <- readRDS("../results/GEX7_BVDJ.rds")

```


## check patient groups in Seurat object

```{r}

table(gex$patient_group, useNA = "always")

#Four non-overlapping groups
#1. SSAB: SSA+/SSB+
#2. SSA+: SSA+/SSB-
#3. SSA-: SSA-/SSB-
#4. CTRL: CTRL


```


# celltypesMain DE tests 
## pseudo bulk DE test between patient subgroups for celltypesMain

```{r }

#https://jef.works/blog/2020/04/06/quickly-creating-pseudobulks/

DGE_list <- list()
for (j in sort(unique(gex$cluster_cellType_manual_main))){

  message(j)
  
  #subset gex to only contain subtype j
  gex.tmp <- gex %>% 
    DietSeurat(dimreducs = "umap_2") %>% 
    tidyseurat::filter(cluster_cellType_manual_main == j)

  #create a model matrix for pseudo bulk generation
  mm <- model.matrix(~ 0 + orig.ident, 
                     data = gex.tmp@meta.data); head(mm); colnames(mm)
  
  #sample x number of cells per condition
  gex.tmp$subgroups <- factor(as.character(gex.tmp$orig.ident))
  sample_size <- table(gex.tmp$subgroups)
  sample_size[sample_size > 100] <- 100
  
  DGE_cells <- lapply(names(sample_size), function(x){ 
    set.seed(1)
    sample(colnames(gex.tmp) [gex.tmp$subgroups == x] , 
           size = sample_size[x], 
           replace = FALSE)
    })
  
  DGE_cells <- unlist(DGE_cells)
  mm <- mm[DGE_cells, ]
  
  #clean up colnames
  colnames(mm) <- gsub("orig.ident", "", colnames(mm))
  
  # create pseudo bulk count matrix
  gex.mm <- gex.tmp@assays$RNA@counts[, DGE_cells] %*% mm; dim(gex.mm)
  gex.mm <- gex.mm[rowSums(gex.mm) != 0, ]; dim(gex.mm)
  
  cellcounts <- colSums(mm); length(cellcounts)
  
  y <- DGEList(gex.mm, 
               sample = gsub("_.*", "", colnames(gex.mm)), 
               group = gsub(".*_", "", colnames(gex.mm)),
               remove.zeros = TRUE); y
  
  y$samples$patient_group <- as.factor(unique(gex.tmp@meta.data[, c("orig.ident", "patient_group")])$patient_group[match(y$samples$samples, unique(gex.tmp@meta.data[, c("orig.ident", "patient_group")])$orig.ident)])
  
  keep <- apply(y$counts, 1, function(x){sum(x >= 5) >= 3}); table(keep)
  y <- y[keep, ]
  y <- calcNormFactors(y)
  
  design <- model.matrix(~ patient_group, data = y$samples); colnames(design)
  colnames(design) <- gsub("patient_group", "", colnames(design))
  
  y_glm <- estimateDisp(y, design)
  fit <- glmQLFit(y_glm, design); colnames(fit$coefficients)
  
  #conserved differences across patient subgroups
  qlm <- glmQLFTest(fit, coef = 2:4)
  de_g <- topTags(qlm, n = nrow(y_glm$counts))$table; head(de_g, 25)
  
  de_g$celltypeMain <- j
  de_g$gene <- rownames(de_g)
  
  DGE_list[[j]] <- de_g
  
  #saveRDS(fit, paste0("../results/DGEfit_patientGroups_celltypeMain_", j,".rds"))
  write.csv(de_g, paste0("../results/DGE_patientGroups_celltypeMain_", j,".csv"), row.names = TRUE)

}

saveRDS(DGE_list, "../results/DGE_B-cells_celltypeMain_patientGroups_listAll.rds")
#DGE_list <- readRDS("../results/DGE_B-cells_celltypeMain_patientGroups_listAll.rds")

DGE_list <- DGE_list[c("Memory_Classical", "Memory_DN", "Memory_IgM", "Naive", "Transitional")]


```


### filter rbind and export csv for all celltypesMain DE tests

```{r}

FDR_cutoff <- 0.25
DGE_list_out <- lapply(DGE_list, function(x){

  
  cat(unique(x$celltypeMain), "\n")
  
  #filter on FDR
  x <- x[x$FDR <= FDR_cutoff, ]
  
  cat(nrow(x), "\n")
  
  #subset GEX to only include cell type
  gex_tmp <- gex %>% 
    DietSeurat(dimreducs = "umap_2") %>% 
    tidyseurat::filter(cluster_cellType_manual_main == unique(x$celltypeMain))
  
  
  #get "pct.exp" and "avg.exp" for filtering
  p <- DotPlot(gex_tmp, features = rownames(x), group.by = "patient_group")
  p <- p$data %>%  pivot_wider(names_from = "id", 
                          values_from = c("pct.exp", "avg.exp"), 
                          id_cols = "features.plot")
  x <- cbind(x, p)
  
  return(x)
  
})



DGE_list_out <- as.data.frame(do.call(rbind, DGE_list_out))

write.csv(DGE_list_out, 
          paste0("../results/DGE_B-cells_celltypeMain_all_FDR", FDR_cutoff, ".csv"), 
          row.names = FALSE)

# DGE_list_out[ ,c("celltypeMain", "gene")] %>% arrange(gene)
# sort(table(DGE_list_out$gene), decreasing = T)


```


### plot Fig2A, filter and plot dotplots for top expressed genes for celltypesMain

```{r}

#filter DE genes
FDR_cutoff <- 0.2
FC_cutoff <- 2
PCTexpr_cutoff <- 20


####---- subset and make plots
DGE_list_p <- lapply(DGE_list, function(a) {

  print(unique(a$celltypeMain))
  
  #filter on FDR
  a <- a[a$FDR <= FDR_cutoff, ]
  
  cat("after FDR: ",nrow(a), "\n")
  
  #subset GEX to only include cell type
  gex_tmp <- gex[a$gene, ] %>% 
    DietSeurat(dimreducs = NULL, graphs = NULL) %>% 
    tidyseurat::filter(cluster_cellType_manual_main == unique(a$celltypeMain))
  
  cat("nrow GEX: ", nrow(gex_tmp@meta.data), "\n")
  
  #get "pct.exp" and "avg.exp" for filtering
  exp <- FetchData(gex_tmp,
                   paste0("rna_", rownames(a)),
                   slot = "counts")
  colnames(exp) <- gsub("rna_", "", colnames(exp))
  exp$patient_group <- gex_tmp$patient_group

  y <- data.frame(matrix(ncol = ncol(exp[, 1:(ncol(exp) - 1)]), nrow = 4))
  colnames(y) <- colnames(exp[, 1:(ncol(exp) - 1)])
  rownames(y) <- levels(gex_tmp$patient_group)
  
  for (i in levels(gex_tmp$patient_group)){
    x <- exp[exp$patient_group == i, 1:(ncol(exp) - 1)]
    y[i, ] <- colSums(x != 0) / nrow(x) * 100
  }
   
  p <- t(y)
  
  a <- cbind(a, p)
  head( a)
  #filter on fold change
  a <- a[apply(a[, 1:3], 1, max) > log2(FC_cutoff) |
         apply(a[, 1:3], 1, min) < log2(1/FC_cutoff), ]
  
  cat("after FC: ",nrow(a), "\n")
  
  #filter on percent expressed
  a <- a[apply(a[, 11:13], 1, max) > PCTexpr_cutoff, ]
  
  cat("after %e: ",nrow(a), "\n")

  return(a)
  
  rm(gex_tmp, p, x, y, exp, a)
  
  }
  
)


####---- plot dotplots to list
ngenes <- 25

p_list <- list()
for (i in sort(unique(gex$cluster_cellType_manual_main))[-c(4, 5, 7)]){

  cat(i,"\n")
  
  a <- as_tibble(DGE_list_p[[i]]) %>% arrange(logFC.SSAB, PValue) %>% pull(gene)
  
  cat(length(a),"\n")
  cat(a,"\n")
  
  if(length(a) < ngenes) {
    p_features <- a
  } else {
    p_features <- a[1:ngenes]
  }
  
  cat(length(p_features),"\n")
  
  #subset gex
  DGE_DATA <- gex[a,] %>% 
  DietSeurat(dimreducs = NULL, graphs = NULL) %>% 
  tidyseurat::filter(cluster_cellType_manual_main == i)

  tmp <-  DotPlot(DGE_DATA,
                         features = p_features,
                         group.by = "patient_group",
                         dot.scale = 5) +
  theme_classic() +
  xlab("") + 
  ylab("") + 
  ggtitle(i) +
  coord_flip() + 
  theme(axis.text.x = element_text(angle = 90, 
                                   hjust = 1, 
                                   vjust = 0.5,
                                   size = 8),
        legend.position ="bottom", 
        legend.box = "horizontal",
        plot.title = element_text(hjust = 0.5,
                                  size = 10),
        panel.grid.major = element_line(colour = "grey", size = rel(0.5)))
  
  p_legend <- get_legend(tmp) %>% ggpubr::as_ggplot()
  
  p_list[[i]] <- tmp + theme(legend.position = "none")
  
  rm(DGE_DATA)

}

p1 <- plot_grid(plotlist = p_list, ncol = 5, labels = NULL, align = "h") + 
           bgcolor("white") + 
           border("white")
  
ggsave2("../results/Fig2A.png", 
        plot_grid(p1, p_legend, ncol = 1, rel_heights = c(8, 1), labels = NULL, align = "h") + 
           bgcolor("white") + 
           border("white"), 
        width = 28, height = 10, unit = "cm")

ggsave2("../results/Fig2A.pdf", 
        plot_grid(p1, p_legend, ncol = 1, rel_heights = c(8, 1), labels = NULL, align = "h") + 
           bgcolor("white") + 
           border("white"), 
        width = 28, height = 10, unit = "cm")


```


### IFN response genes in DE list ?
```{r}

DGE_list_p_df <- as.data.frame(do.call(rbind, DGE_list_p))
DGE_list_p_df[ ,c("celltypeMain", "gene")] %>% arrange(gene)


# get hallmark gene sets
HM <- getGenesets("/Users/gusarv/Downloads/MSigDB_Hallmark_2020.txt")

IFN <- unique(c(HM$`Interferon Alpha Response`, 
                HM$`Interferon Gamma Response`))
  
DGE_list_p_df_ifn <- DGE_list_p_df[DGE_list_p_df$gene %in% IFN, ]
DGE_list_p_df_ifn[ ,c("celltypeMain", "gene")] %>% arrange(gene)


```



## enrichment tests for celltypesMain (MSigDB_Hallmark_2020 and GO_Biological_Process_2021)

```{r}

p_listH <- list()
p_listBP <- list()
gse_listH <- list()
gse_listBP <- list()

for (i in names(DGE_list)){
  
  go_cluster <- enrichr(genes = DGE_list[[i]]$gene[1:50], 
                        databases = c("GO_Biological_Process_2021", 
                                      "MSigDB_Hallmark_2020"))
  
  go_clusterBP <- go_cluster$GO_Biological_Process_2021
  go_clusterBP <- go_clusterBP[order(go_clusterBP$P.value), ]
  go_clusterBP$logP <- -log10(go_clusterBP$P.value)
  go_clusterBP$logPadj <- -log10(go_clusterBP$Adjusted.P.value)
  gse_listBP[[i]] <- go_clusterBP
  
  go_clusterH <- go_cluster$MSigDB_Hallmark_2020
  go_clusterH <- go_clusterH[order(go_clusterH$P.value), ]
  go_clusterH$logP <- -log10(go_clusterH$P.value)
  go_clusterH$logPadj <- -log10(go_clusterH$Adjusted.P.value)
  gse_listH[[i]] <- go_clusterH
  
  p_listH[[i]] <- ggplot(go_clusterH[1:6,], 
                         aes(x = reorder(Term, logP), 
                             y = logPadj)) +
    geom_col(position = "dodge") +
    theme_classic() + 
    coord_flip() +
    theme(axis.text.x = element_text(angle = 0, 
                                     hjust = 0.5, 
                                     vjust = 0.5,
                                     size = 12,
                                     color = "black"),
          axis.text.y = element_text(angle = 0, 
                                     hjust = 1, 
                                     vjust = 0.5,
                                     size = 12,
                                     color = "black"),
          strip.background = element_blank(), 
          strip.text.y = element_blank(),
          legend.position = "none",
          plot.margin = unit(c(0.2,0.2,0.2,0.2), "cm")) +
      ylab(expression('-log10('~italic(P)~' adj)')) +
      xlab("") +
      ggtitle(paste0(i)) +
      geom_hline(aes(yintercept = 2), color = "black", linetype = "dashed")
  
  p_listBP[[i]] <- ggplot(go_clusterBP[1:15,], 
                          aes(x = reorder(Term, logP), 
                              y = logPadj)) +
    geom_col(position = "dodge") +
    theme_classic() + 
    coord_flip() +
    theme(axis.text.x = element_text(angle = 0, 
                                     hjust = 0.5, 
                                     vjust = 0.5,
                                     size = 12,
                                     color = "black"),
          axis.text.y = element_text(angle = 0, 
                                     hjust = 1, 
                                     vjust = 0.5,
                                     size = 12,
                                     color = "black"),
          strip.background = element_blank(), 
          strip.text.y = element_blank(),
          legend.position = "none",
          plot.margin = unit(c(0.2, 0.2, 0.2, 0.2), "cm")) +
      scale_y_continuous(breaks = seq(0, 20, 5), limits = c(0, 20)) +
      ylab(expression('-log10('~italic(P)~' adj)')) +
      xlab("") +
      ggtitle(paste0(i)) +
      geom_hline(aes(yintercept = 2), color = "black", linetype = "dashed")
  
}

# write.csv(do.call(rbind.data.frame, gse_listH),
#           "../results/DGE_B-cells_GSE_celltypeMain_hallmarks.csv", 
#           row.names = FALSE)
# 
# write.csv(do.call(rbind.data.frame, gse_listBP),
#           "../results/DGE_B-cells_GSE_celltypeMain_GOBP.csv", 
#           row.names = FALSE)


```


### plot Fig2B, GSEA celltypesMain (MSigDB_Hallmark_2020)

```{r}

ggsave2(paste0("../results/Fig2B.png"),
          plot_grid(plotlist = p_listH, 
                    ncol = 5, 
                    labels = NULL, 
                    align = "vh" ) + 
          bgcolor("white") + 
          border("white"),
          width = 52, height = 8, unit = "cm")

ggsave2(paste0("../results/Fig2B.pdf"),
          plot_grid(plotlist = p_listH, 
                    ncol = 5, 
                    labels = NULL, 
                    align = "vh" ) + 
          bgcolor("white") + 
          border("white"),
          width = 52, height = 8, unit = "cm")


```


### plot SFig6, GSEA celltypesMain (GO_Biological_Process_2021)

```{r}

ggsave2(paste0("../results/SFig6.png"),
          plot_grid(plotlist = p_listBP, 
                    ncol = 2, 
                    labels = NULL, 
                    align = "vh" ) + 
          bgcolor("white") + 
          border("white"),
          width = 50, height = 35, unit = "cm")

ggsave2(paste0("../results/SFig6.pdf"),
          plot_grid(plotlist = p_listBP, 
                    ncol = 2, 
                    labels = NULL, 
                    align = "vh" ) + 
          bgcolor("white") + 
          border("white"),
          width = 50, height = 35, unit = "cm")


```


#celltypesFine DE tests 
## pseudo bulk DE test between patient subgroups for celltypesFine

```{r}

#https://jef.works/blog/2020/04/06/quickly-creating-pseudobulks/

DGE_list <- list()
for (j in sort(unique(gex$cluster_cellType_manual_fine))){

  message(j)
  
  #subset gex to only contain subtype j
  gex.tmp <- gex %>% 
    DietSeurat(dimreducs = "umap_2") %>% 
    tidyseurat::filter(orig.ident != "P024", 
                       cluster_cellType_manual_fine == j)

  #create a model matrix for pseudo bulk generation
  mm <- model.matrix(~ 0 + orig.ident, 
                     data = gex.tmp@meta.data); head(mm); colnames(mm)
  
  #sample x number of cells per condition
  gex.tmp$subgroups <- factor(as.character(gex.tmp$orig.ident))
  sample_size <- table(gex.tmp$subgroups)
  sample_size[sample_size > 100] <- 100
  
  DGE_cells <- lapply(names(sample_size), function(x){ 
    set.seed(1)
    sample(colnames(gex.tmp) [gex.tmp$subgroups == x] , 
           size = sample_size[x], 
           replace = FALSE)
    })
  
  DGE_cells <- unlist(DGE_cells)
  mm <- mm[DGE_cells, ]
  
  #clean up colnames
  colnames(mm) <- gsub("orig.ident", "", colnames(mm))
  
  # create pseudo bulk count matrix
  gex.mm <- gex.tmp@assays$RNA@counts[, DGE_cells] %*% mm; dim(gex.mm)
  gex.mm <- gex.mm[rowSums(gex.mm) != 0, ]; dim(gex.mm)
  
  cellcounts <- colSums(mm); length(cellcounts)
  
  y <- DGEList(gex.mm, 
               sample = gsub("_.*", "", colnames(gex.mm)), 
               group = gsub(".*_", "", colnames(gex.mm)),
               remove.zeros = TRUE); y
  
  y$samples$patient_group <- as.factor(unique(gex.tmp@meta.data[, c("orig.ident", "patient_group")])$patient_group[match(y$samples$samples, unique(gex.tmp@meta.data[, c("orig.ident", "patient_group")])$orig.ident)])
  
  keep <- apply(y$counts, 1, function(x){sum(x >= 5) >= 3}); table(keep)
  y <- y[keep, ]
  y <- calcNormFactors(y)
  
  design <- model.matrix(~ patient_group, data = y$samples); colnames(design)
  colnames(design) <- gsub("patient_group", "", colnames(design))
  
  y_glm <- estimateDisp(y, design)
  fit <- glmQLFit(y_glm, design); colnames(fit$coefficients)
  
  #conserved differences across patient subgroups
  qlm <- glmQLFTest(fit, coef = 2:4)
  de_g <- topTags(qlm, n = nrow(y_glm$counts))$table; head(de_g, 25)
  
  de_g$celltypeFine <- j
  de_g$gene <- rownames(de_g)
  
  DGE_list[[j]] <- de_g
  
  saveRDS(fit, paste0("../results/DGEfit_patientGroups_celltypeFine_", j,".rds"))
  write.csv(de_g, paste0("../results/DGE_patientGroups_celltypeFine_", j,".csv"), row.names = TRUE)

}

saveRDS(DGE_list, "../results/DGE_B-cells_celltypeFine_patientGroups_listAll.rds")
#DGE_list <- readRDS("../results/DGE_B-cells_celltypeFine_patientGroups_listAll.rds")
#lapply(DGE_list, function(x){head(x)})


```


## filter rbind and export csv for all celltypeFine DE tests

```{r}

FDR_cutoff <- 0.25
DGE_list_out <- lapply(DGE_list, function(x){

  
  cat(unique(x$celltypeFine), "\n")
  
  #filter on FDR
  x <- x[x$FDR <= FDR_cutoff, ]
  
  cat(nrow(x), "\n")
  
  gex_tmp <- gex %>% 
    DietSeurat(dimreducs = "umap_2") %>% 
    tidyseurat::filter(cluster_cellType_manual_fine == unique(x$celltypeFine))
  
  #get "pct.exp" and "avg.exp" for filtering
  p <- DotPlot(gex_tmp, features = rownames(x), group.by = "patient_group")
  p <- p$data %>%  pivot_wider(names_from = "id", 
                          values_from = c("pct.exp", "avg.exp"), 
                          id_cols = "features.plot")
  x <- cbind(x, p)
  
  return(x)
  
})


DGE_list_out <- as.data.frame(do.call(rbind, DGE_list_out))

write.csv(DGE_list_out, 
          paste0("../results/DGE_B-cells_celltypeFine_all_FDR", FDR_cutoff, ".csv"), 
          row.names = FALSE)


```


### plot SFig4, filter and plot dotplots for top expressed genes celltypeFine, Supplementary Figure 5

```{r}

#filter DE genes
FDR_cutoff <- 0.2
FC_cutoff <- 2
PCTexpr_cutoff <- 20

plotlist <- c("Memory_DN1_ID3", "Memory_DN2", "Memory_DN2_CXCR3", "Memory_DN2_ITGAX", "Memory_DN4",
              "Memory_Classical", "Memory_IgM", "Memory_IgM_ALOX5", "Memory_IgM_CD1C", 
              "Naive", "Naive_IFN", "Naive_Transitional")
DGE_list <- DGE_list[match(plotlist, names(DGE_list))]


####---- subset and make plots
DGE_list_p <- lapply(DGE_list, function(a) {
  
  
  print(unique(a$celltypeFine))
  
  #filter on FDR
  a <- a[a$FDR <= FDR_cutoff, ]
  
  cat("after FDR: ",nrow(a), "\n")
  
  gex_tmp <- gex[a$gene, ] %>% 
    DietSeurat(dimreducs = NULL, graphs = NULL) %>% 
    tidyseurat::filter(cluster_cellType_manual_fine == unique(a$celltypeFine))
  
 #----- get "pct.exp" and "avg.exp" for filtering
  exp <- FetchData(gex_tmp,
                   rownames(a),
                   slot = "counts")
                   
  exp$patient_group <- gex_tmp$patient_group

  y <- data.frame(matrix(ncol = ncol(exp[, 1:(ncol(exp) - 1)]), nrow = 4))
  colnames(y) <- colnames(exp[, 1:(ncol(exp) - 1)])
  rownames(y) <- levels(gex_tmp$patient_group)
  
  for (i in levels(gex_tmp$patient_group)){
    x <- exp[exp$patient_group == i, 1:(ncol(exp) - 1)]
    y[i, ] <- colSums(x != 0) / nrow(x) * 100
  }
   
  p <- t(y)
  #-----
  
  a <- cbind(a, p)
  
  #filter on fold change
  a <- a[apply(a[, 1:3], 1, max) > log2(FC_cutoff) |
         apply(a[, 1:3], 1, min) < log2(1/FC_cutoff), ]
  
  cat("after FC: ",nrow(a), "\n")
  
  #filter on percent expressed
  a <- a[apply(a[, 11:13], 1, max) > PCTexpr_cutoff, ]
  
  cat("after %e: ",nrow(a), "\n")

  return(a)
  
  rm(gex_tmp)
  rm(p)
  rm(y)
  rm(exp)
  rm(a)
  
   }
  )


####---- plot dotplots to list
ngenes <- 20

p_list <- list()
for (i in sort(unique(plotlist))){

  cat(i,"\n")
  
  a <- as_tibble(DGE_list_p[[i]]) %>% arrange(logFC.SSAB, PValue) %>% pull(gene)
  
  cat(length(a),"\n")
  cat(a,"\n")
  
  if(length(a) < ngenes) {
    p_features <- a
  } else {
    p_features <- a[1:ngenes]
  }
  
  cat(length(p_features),"\n")
  
  #subset gex
  DGE_DATA <- gex[a,] %>% 
  DietSeurat(dimreducs = NULL, graphs = NULL) %>% 
  tidyseurat::filter(orig.ident != "P024", 
                     cluster_cellType_manual_fine == i)

  tmp <- DotPlot(DGE_DATA, 
        features = p_features, 
        group.by = "patient_group",
        dot.scale = 4) +
  theme_classic() +
  xlab("") + 
  ylab("") + 
  ggtitle(i) +
  coord_flip() + 
  theme(axis.text.x = element_text(angle = 90, 
                                   hjust = 1, 
                                   vjust = 0.5,
                                   size = 8),
        legend.position ="bottom", 
        legend.box = "horizontal",
        plot.title = element_text(hjust = 0.5),
        panel.grid.major = element_line(colour = "grey", size = rel(0.5)))

  #extract legend
  p_legend <- get_legend(tmp) %>% ggpubr::as_ggplot()
  
  #remove legend
  p_list[[i]] <- tmp + theme(legend.position = "none")
  
  rm(DGE_DATA)
  
}

p1 <- plot_grid(plotlist = p_list, ncol = 3, labels = NULL, align = "h") + 
           bgcolor("white") + 
           border("white")

ggsave2("../results/SFig4.png", 
        plot_grid(p1, p_legend, ncol = 1, labels = NULL, rel_heights = c(20, 1), align = "hv") + 
           bgcolor("white") + 
           border("white"), 
        width = 22, height = 31, unit = "cm")

ggsave2("../results/SFig4.pdf", 
        plot_grid(p1, p_legend, ncol = 1, labels = NULL, rel_heights = c(20, 1), align = "hv") + 
           bgcolor("white") + 
           border("white"), 
        width = 22, height = 31, unit = "cm")


```


## enrichment tests for celltypeFine (MSigDB_Hallmark_2020 and GO_Biological_Process_2021)

```{r}

p_listH <- list()
p_listBP <- list()
gse_listH <- list()
gse_listBP <- list()

plotlist <- c("Memory_DN1_ID3", "Memory_DN2", "Memory_DN2_CXCR3", "Memory_DN2_ITGAX", "Memory_DN4",
              "Memory_Classical", "Memory_IgM", "Memory_IgM_ALOX5", "Memory_IgM_CD1C", 
              "Naive", "Naive_IFN", "Naive_Transitional")

for (i in plotlist){
  
  go_cluster <- enrichr(genes = DGE_list[[i]]$gene[1:50], 
                        databases = c("GO_Biological_Process_2021", 
                                      "MSigDB_Hallmark_2020"))
  
  go_clusterBP <- go_cluster$GO_Biological_Process_2021
  go_clusterBP <- go_clusterBP[order(go_clusterBP$P.value), ]
  go_clusterBP$logP <- -log10(go_clusterBP$P.value)
  go_clusterBP$logPadj <- -log10(go_clusterBP$Adjusted.P.value)
  gse_listBP[[i]] <- go_clusterBP
  
  go_clusterH <- go_cluster$MSigDB_Hallmark_2020
  go_clusterH <- go_clusterH[order(go_clusterH$P.value), ]
  go_clusterH$logP <- -log10(go_clusterH$P.value)
  go_clusterH$logPadj <- -log10(go_clusterH$Adjusted.P.value)
  gse_listH[[i]] <- go_clusterH
  
  
  p_listH[[i]] <- ggplot(go_clusterH[1:6,], aes(x = reorder(Term, logP), 
                                              y = logPadj)) +
    geom_col(position = "dodge") +
    theme_classic() + 
    coord_flip() +
    theme(axis.text.x = element_text(angle = 0, 
                                     hjust = 0.5, 
                                     vjust = 0.5,
                                     size = 12,
                                     color = "black"),
          axis.text.y = element_text(angle = 0, 
                                     hjust = 1, 
                                     vjust = 0.5,
                                     size = 12,
                                     color = "black"),
          plot.title = element_text(hjust = 0),
          strip.background = element_blank(), 
          strip.text.y = element_blank(),
          legend.position = "none",
          plot.margin = unit(c(0.2,0.2,0.2,0.2), "cm")) +
      ylab(expression('-log10('~italic(P)~' adj)')) +
      xlab("") +
      scale_y_continuous(breaks = seq(0, 25, 5), limits = c(0, 25)) +
      ggtitle(paste0(i)) +
      geom_hline(aes(yintercept = 2), color = "black", linetype = "dashed")
  
  p_listBP[[i]] <- ggplot(go_clusterBP[1:15,], aes(x = reorder(Term, logP), 
                                              y = logPadj)) +
    geom_col(position = "dodge") +
    theme_classic() + 
    coord_flip() +
    theme(axis.text.x = element_text(angle = 0, 
                                     hjust = 0.5, 
                                     vjust = 0.5,
                                     size = 9,
                                     color = "black"),
          axis.text.y = element_text(angle = 0, 
                                     hjust = 1, 
                                     vjust = 0.5,
                                     size = 8,
                                     color = "black"),
          plot.title = element_text(hjust = 0),
          strip.background = element_blank(), 
          strip.text.y = element_blank(),
          legend.position = "none",
          plot.margin = unit(c(0.2,0.2,0.2,0.2), "cm")) +
      ylab(expression('-log10('~italic(P)~' adj)')) +
      xlab("") +
      scale_y_continuous(breaks = seq(0, 25, 5), limits = c(0, 25)) +
      ggtitle(paste0(i)) +
      geom_hline(aes(yintercept = 2), color = "black", linetype = "dashed")
  
}


# save enrichment output
write.csv(do.call(rbind.data.frame, gse_listH),
          "../results/DGE_B-cells_GSE_celltypeFine_hallmarks.csv", 
          row.names = FALSE)

write.csv(do.call(rbind.data.frame, gse_listBP),
          "../results/DGE_B-cells_GSE_celltypeFine_GOBP.csv", 
          row.names = FALSE)


```


### plot SFig5, GSEA celltypesFine (MSigDB_Hallmark_2020)

```{r}

ggsave2(paste0("../results/SFig5.png"),
          plot_grid(plotlist = p_listH, 
                    ncol = 2, 
                    labels = NULL, 
                    align = "vh" )  +
          bgcolor("white") + 
          border("white"),
          width = 35, height = 30, unit = "cm")

ggsave2(paste0("../results/SFig5.pdf"),
          plot_grid(plotlist = p_listH, 
                    ncol = 2, 
                    labels = NULL, 
                    align = "vh" )  +
          bgcolor("white") + 
          border("white"),
          width = 35, height = 30, unit = "cm")


```


### plot SFig7, GSEA celltypesFine (GO_Biological_Process_2021)

```{r}
ggsave2(paste0("../results/SFig7.png"),
          plot_grid(plotlist = p_listBP, 
                    ncol = 2, 
                    labels = NULL, 
                    align = "vh" )  +
          bgcolor("white") + 
          border("white"),
          width = 47, height = 35, unit = "cm")

ggsave2(paste0("../results/SFig7.pdf"),
          plot_grid(plotlist = p_listBP, 
                    ncol = 2, 
                    labels = NULL, 
                    align = "vh" )  +
          bgcolor("white") + 
          border("white"),
          width = 47, height = 35, unit = "cm")


```


## print sessionInfo()

```{r}

sessionInfo()

```
