==> code/09-variablePCA.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# PCA per sample variables {#PCA}

## load packages

```{r}

source("./00a_dependencies.R")

```


## load Seurat object and reduce size for local processing

```{r}

gex.slim <- readRDS("../results/GEX7_BVDJ.rds")
gex.slim <- DietSeurat(gex.slim, counts = TRUE, dimreducs = NULL, graphs = NULL, scale.data = FALSE)
gex.slim <- gex.slim[, (gex.slim$cluster_cellType_manual_fine == "Memory_DN1_ID3" |
                        gex.slim$cluster_cellType_manual_fine == "Memory_DN2" |
                        gex.slim$cluster_cellType_manual_fine == "Memory_DN2_CXCR3" |
                        gex.slim$cluster_cellType_manual_fine == "Memory_DN2_ITGAX" |
                        gex.slim$cluster_cellType_manual_fine == "Memory_DN4" |
                        gex.slim$cluster_cellType_manual_fine == "Memory_Classical" |
                        gex.slim$cluster_cellType_manual_fine == "Memory_IgM" |
                        gex.slim$cluster_cellType_manual_fine == "Memory_IgM_ALOX5" |
                        gex.slim$cluster_cellType_manual_fine == "Memory_IgM_CD1C" |
                        gex.slim$cluster_cellType_manual_fine == "Naive" |
                        gex.slim$cluster_cellType_manual_fine == "Naive_IFN" |
                        gex.slim$cluster_cellType_manual_fine == "Naive_Transitional")]


```


## load patient Meta Data

```{r}

source("./00b_pheno.R")

pheno <- pheno[-28,] # remove P024

#make binary data columns numerical binary
pheno[pheno == "NO"] <- 0
pheno[pheno == "YES"] <- 1
pheno[pheno == "MD"] <- NA
pheno[, 2:(ncol(pheno)-2)] <- sapply(pheno[, 2:(ncol(pheno) - 2)], as.numeric)
colnames(pheno)

#remove variables only present in one sample
pheno <- pheno[, c(1, (1 + which(colSums(pheno[,2:29], na.rm = TRUE) > 2 )), 36)]
pheno <- pheno[, !(names(pheno) %in% c("Treatment_at_B_cell_sampling_HCQ", "arthralgia_firstSymptom"))]
colnames(pheno)

pheno$group <- as.character(pheno$patient_group)
pheno$group[5:nrow(pheno)] <- "pSS"
pheno$group <- as.factor(pheno$group)

pheno <- pheno[, c(1, ncol(pheno), ncol(pheno) - 1, 2:(ncol(pheno) - 2)) ]


```


## 1. celltype frequencies per patient

```{r}

cells_main <- dplyr::select(gex.slim@meta.data,
                       orig.ident,
                       cluster_cellType_manual_main) %>%
  add_count(orig.ident,
            cluster_cellType_manual_main,
            .drop = FALSE) %>%
  distinct() %>%
  group_by(orig.ident) %>%
  mutate(per =  prop.table(n)*100) %>%
  drop_na() %>%
  dplyr::select(orig.ident,
                cluster_cellType_manual_main,
                per) %>%
  pivot_wider(names_from = cluster_cellType_manual_main,
              values_from = per)

colnames(cells_main)[2:ncol(cells_main)] <- paste0(colnames(cells_main)[2:ncol(cells_main)], "_%")

#saveRDS(cells_main, file = "../results/forPCA_cellsMain.rds")
cells_main <- readRDS("../results/forPCA_cellsMain.rds")


```


## 2. VDJ percentages per celltype

```{r}

# IGHV
IGHV <- dplyr::select(gex.slim@meta.data,
                           orig.ident,
                           IGHV_gene) %>%
  add_count(orig.ident,
            IGHV_gene,
            .drop = FALSE) %>%
  distinct() %>%
  group_by(orig.ident) %>%
  mutate(per =  prop.table(n)*100) %>%
  drop_na() %>%
  dplyr::select(orig.ident, IGHV_gene, per) %>%
  pivot_wider(
    names_from = IGHV_gene,
    values_from = per,
    values_fill = 0
  )

colnames(IGHV)[2:ncol(IGHV)] <-
  paste0(colnames(IGHV)[2:ncol(IGHV)], "_%")


# IGHD
IGHD <- dplyr::select(gex.slim@meta.data,
                           orig.ident,
                           IGHD_gene) %>%
  add_count(orig.ident,
            IGHD_gene,
            .drop = FALSE) %>%
  distinct() %>%
  group_by(orig.ident) %>%
  mutate(per =  prop.table(n)*100) %>%
  drop_na() %>%
  dplyr::select(orig.ident, IGHD_gene, per) %>%
  pivot_wider(
    names_from = c(IGHD_gene),
    values_from = per,
    values_fill = 0
  )

colnames(IGHD)[2:ncol(IGHD)] <-
  paste0(colnames(IGHD)[2:ncol(IGHD)], "_%")


# IGHJ
IGHJ <- dplyr::select(gex.slim@meta.data,
                           orig.ident,
                           IGHJ_gene) %>%
  add_count(orig.ident,
            IGHJ_gene,
            .drop = FALSE) %>%
  distinct() %>%
  group_by(orig.ident) %>%
  mutate(per =  prop.table(n)*100) %>%
  drop_na() %>%
  dplyr::select(orig.ident, IGHJ_gene, per) %>%
  pivot_wider(
    names_from = IGHJ_gene,
    values_from = per,
    values_fill = 0
  )

colnames(IGHJ)[2:ncol(IGHJ)] <-
  paste0(colnames(IGHJ)[2:ncol(IGHJ)], "_%")


# # IGHC gene per celltypeMine
c_gene_10x_main <- dplyr::select(gex.slim@meta.data,
                                 orig.ident,
                                 cluster_cellType_manual_main,
                                 c_gene_10x) %>%
  add_count(orig.ident,
            cluster_cellType_manual_main,
            c_gene_10x, .drop = FALSE) %>%
  distinct() %>%
  group_by(orig.ident, cluster_cellType_manual_main) %>%
  mutate(per =  prop.table(n)*100) %>%
  drop_na() %>%
  dplyr::select(orig.ident, cluster_cellType_manual_main, c_gene_10x, per) %>%
  pivot_wider(
    names_from = c(c_gene_10x, cluster_cellType_manual_main),
    values_from = per,
    values_fill = 0
  )

colnames(c_gene_10x_main)[2:ncol(c_gene_10x_main)] <-
  paste0(colnames(c_gene_10x_main)[2:ncol(c_gene_10x_main)], "_%")


```


## 3. IFN score

```{r}

# get hallmark gene sets
HM <- getGenesets("/Users/gusarv/Downloads/MSigDB_Hallmark_2020.txt")

IFN <- unique(c(HM$`Interferon Alpha Response`, 
                HM$`Interferon Gamma Response`))
sort(IFN)
#cluster_cellType_manual_main
ALL_expr <- AggregateExpression(gex.slim,
                                slot = "counts",
                                group.by = c("cluster_cellType_manual_main", "orig.ident"))

ALL_expr <- as.data.frame(ALL_expr$RNA)
IFN_expr <- ALL_expr[IFN, ]
#ALL_expr <- ALL_expr[!(rownames(ALL_expr) %in% IFN), ]

IFN_main <- as.data.frame(colSums(IFN_expr, na.rm = TRUE) / colSums(ALL_expr, na.rm = TRUE) * 100)

colnames(IFN_main)[1] <- "IFN_score"
IFN_main$cluster_cellType_manual_main <- substr(rownames(IFN_main), 1, nchar(rownames(IFN_main)) - 5)
IFN_main$orig.ident <- substr(rownames(IFN_main), nchar(rownames(IFN_main)) - 3, nchar(rownames(IFN_main)))

IFN_main <- pivot_wider(IFN_main,
                        names_from = cluster_cellType_manual_main,
                        values_from = IFN_score)

colnames(IFN_main)[2:ncol(IFN_main)] <- paste0(colnames(IFN_main)[2:ncol(IFN_main)], "_IFNscore")

#saveRDS(IFN_main, file = "../results/forPCA_IFN_main.rds")
IFN_main <- readRDS("../results/forPCA_IFN_main.rds")


```


## 4. mutated VDJ bp per celltype 

```{r}

mu_main <- gex.slim@meta.data %>%
  select(orig.ident, cluster_cellType_manual_main, mu_freq_total, c_gene_10x) %>%
  drop_na() %>%
  group_by(orig.ident, cluster_cellType_manual_main) %>%
  summarise(mu_mean_main = mean(mu_freq_total * 100)) %>%
  pivot_wider(names_from = cluster_cellType_manual_main,
              values_from = mu_mean_main,
              values_fill = 0)

colnames(mu_main)[2:ncol(mu_main)] <- paste0(colnames(mu_main)[2:ncol(mu_main)], "_VDJmutation")


```


## 5. clonotypes % 

```{r}

# clonotypes for cell type main
clono_main <- gex.slim@meta.data %>%
  select(orig.ident, cluster_cellType_manual_main, cloneType_airr_short, c_gene_10x) %>%
  drop_na() %>%
  select(orig.ident, cluster_cellType_manual_main, cloneType_airr_short)  %>%
  add_count(orig.ident,
            cluster_cellType_manual_main,
            cloneType_airr_short,
            .drop = FALSE) %>%
  distinct() %>%
  group_by(orig.ident, cluster_cellType_manual_main) %>%
  mutate(per =  prop.table(n)*100) %>%
  dplyr::select(orig.ident,
                cluster_cellType_manual_main,
                cloneType_airr_short,
                per) %>%
  pivot_wider(names_from = c(cluster_cellType_manual_main, cloneType_airr_short),
              values_from = per,
              values_fill = 0)

colnames(clono_main)[2:ncol(clono_main)] <- paste0(colnames(clono_main)[2:ncol(clono_main)], "_clonotypes_%")


```


## concatenate vars object

```{r}

var_df <- 
  full_join(cells_main, IGHV, by = "orig.ident") %>%
  full_join(IGHD, by = "orig.ident") %>%
  full_join(IGHJ, by = "orig.ident") %>%
  full_join(c_gene_10x_main, by = "orig.ident") %>%
  full_join(IFN_main, by = "orig.ident") %>%
  full_join(mu_main, by = "orig.ident") %>%
  full_join(clono_main, by = "orig.ident")

var_df[is.na(var_df)] <- 0

#saveRDS(var_df, "../results/forPCA_variables_per_patient.rds")
#var_df <- readRDS("../results/forPCA_variables_per_patient.rds")

```


## color df for plotting

```{r}
colors_df <- data.frame(names = colnames(var_df),
                        category = NA,
                        color = NA)
rownames(colors_df) <- colors_df$names


colors_df[colors_df$names %in% colnames(cells_main)[2:ncol(cells_main)], "category"] <- "cellType"
colors_df[colors_df$names %in% c(colnames(IGHV)[2:ncol(IGHV)],
                                 colnames(IGHD)[2:ncol(IGHD)],
                                 colnames(IGHJ)[2:ncol(IGHJ)],
                                 colnames(c_gene_10x_main)[2:ncol(c_gene_10x_main)]), "category"] <- "VDJ_gene"
colors_df[colors_df$names %in% colnames(c_gene_10x_main)[2:ncol(c_gene_10x_main)], "category"] <- "VDJ_c_gene"
colors_df[colors_df$names %in% colnames(IFN_main)[2:ncol(IFN_main)], "category"] <- "IFN_score"
colors_df[colors_df$names %in% colnames(mu_main)[2:ncol(mu_main)], "category"] <- "VDJ_mutation"
colors_df[colors_df$names %in% colnames(clono_main)[2:ncol(clono_main)], "category"] <- "ClonoType"

colors_df$category <- factor(colors_df$category, levels = unique(colors_df$category)[!is.na(unique(colors_df$category))])
```



## PCA (nipals)

```{r}

#selecting variables with significant differences between pSS samples and CTRLs
m <- apply(var_df[, 2:ncol(var_df)], 2, function(x){glm(x ~pheno$group)} ) 
m <- lapply(m, summary)
m <- as.data.frame(t(sapply(m, function(x){x$coefficients[2, 4]})))

sell <- names(m)[m < 0.1]
length(sell) #[1] 30

pc_df <- as.data.frame(var_df)
rownames(pc_df) <- var_df$orig.ident

# nipals package
pc_out <- nipals::nipals(pc_df[, sell], 
                       ncomp = 4, 
                       center = TRUE, 
                       scale = TRUE, 
                       fitted = FALSE, 
                       maxiter = 500, 
                       tol = 1e-9, 
                       gramschmidt = FALSE)


#saveRDS(pc_out, "../results/PCA_out_var_df.rds")
pc_out <- readRDS("../results/PCA_out_var_df.rds")


```


## plot PCA-out to list

```{r}

pc_out_sc <- as.data.frame(pc_out$scores)
pc_out_sc$orig.ident <- rownames(pc_out_sc)
pc_out_sc$patient_group <- pheno$patient_group

pc_out_ld <- as.data.frame(pc_out$loadings)
pc_out_ld$variable <- rownames(pc_out_ld)
pc_out_ld$category <- colors_df$category[match(pc_out_ld$var, colors_df$names)]


#plot scores    
  p1.1 <- ggplot(pc_out_sc) + 
    geom_text_repel(size = 3,
                     aes(x = PC1, 
                         y = PC2, 
                         label = orig.ident, 
                         color = patient_group), 
                    show.legend = FALSE) +
    xlab(paste0("PC1", " (", round(pc_out$eig[1], digits = 1), "%)")) +
    ylab(paste0("PC2", " (", round(pc_out$eig[2], digits = 1), "%)")) +
    ggtitle("Sample scores") + 
    geom_point(aes(x = PC1, y = PC2, color = patient_group)) +
    theme_classic(base_size = 11) +
    geom_hline(yintercept = 0, linetype = 2, alpha = 0.4, size = 0.5) + 
    geom_vline(xintercept = 0, linetype = 2, alpha = 0.4, size = 0.5) +
    theme(legend.position = "bottom",
          legend.title = element_blank())
  
  p1.2 <- ggplot(pc_out_sc) + 
    geom_text_repel(size = 3,
                     aes(x = PC3, 
                         y = PC4, 
                         label = orig.ident, 
                         color = patient_group),
                    show.legend = FALSE) +
    xlab(paste0("PC3", " (", round(pc_out$eig[3], digits = 1), "%)")) +
    ylab(paste0("PC4", " (", round(pc_out$eig[4], digits = 1), "%)")) +
    ggtitle("Patient group") +
    geom_point(aes(x = PC3, y = PC4, color = patient_group)) +
    geom_hline(yintercept = 0, linetype = 2, alpha = 0.4, size = 0.5) + 
    geom_vline(xintercept = 0, linetype = 2, alpha = 0.4, size = 0.5) +
    theme_classic(base_size = 11) +
    theme(legend.position = "bottom",
          legend.title=element_blank())
  
  p3.1 <- ggplot(pc_out_ld) + 
    geom_text_repel(size = 3,
                    aes(x = PC1, 
                        y = PC2, 
                        label = variable, 
                        color = category),
                    show.legend = FALSE) +
    # xlab(paste0("PC1", " (", round(pc_out$eig[1], digits = 1), "%)")) +
    # ylab(paste0("PC2", " (", round(pc_out$eig[2], digits = 1), "%)")) +
    xlab(paste0("PC1")) +
    ylab(paste0("PC2")) +
    ggtitle("Variable Loadings") +
    geom_point(aes(x = PC1, y = PC2, color = category)) +
    theme_classic(base_size = 11) +
    scale_color_brewer(palette = "Dark2", name = "") +
    geom_hline(yintercept = 0, linetype = 2, alpha = 0.4, size = 0.5) + 
    geom_vline(xintercept = 0, linetype = 2, alpha = 0.4, size = 0.5) +
    theme(legend.position = "bottom",
          legend.margin = margin(),
          legend.spacing.x = grid::unit(-0.5, "mm")
          )


```


## correlation with pheno data

```{r}

# merge PCA output with input and phenmo data
pc_out_full <- full_join(pc_out_sc, pheno, by = c("orig.ident", "patient_group")) %>%
  full_join(pc_df, by = "orig.ident") 


# test for correlations between principal components and phenotype data for pSS samples
p_pheno1 <- data.frame(row.names = colnames(pc_out_full[, 8:25]))
for (i in c("PC1", "PC2")){
  m <- apply(pc_out_full[!pc_out_full$group == "CTRL", rownames(p_pheno1)], 
             2, 
             function(x){glm(pc_out_full[!pc_out_full$group == "CTRL", i] ~ x)} ) 
  m <- lapply(m, summary)
  p_pheno1[, i] <- c(t(sapply(m, function(x){x$coefficients[2, 4]})))
}

p_pheno1$group <- as.factor(c(rep("age", times = 2),
                   rep("auto-Ab", times = 7),
                   rep("other", times = 9)))

p_pheno1 <- p_pheno1[!(row.names(p_pheno1) %in% c("SSA_Ro52", "SSA_Ro60", "SSA_and_SSB")), ]


p5.1 <- ggplot(p_pheno1[1:15, ], aes(x = -log10(PC1),
                                    y = -log10(PC2),
                                    color = group)) +
  geom_point() +
  geom_text_repel(size = 2,
                  aes(label = rownames(p_pheno1[1:15, ])), 
                  show.legend = FALSE) +
  geom_hline(yintercept = -log10(0.05), linetype = 2, alpha = 0.4, size = 0.5) + 
  geom_vline(xintercept = -log10(0.05), linetype = 2, alpha = 0.4, size = 0.5) +
  ylab(expression(paste("-log10(", italic("P"), ") PC2"))) +
  xlab(expression(paste("-log10(", italic("P"), ") PC1"))) +
  ggtitle("glm(PC ~ variable)") +
  ylim(0, 3) +
  xlim(0, 6.5) +
  scale_color_brewer(palette="Dark2") +
  theme_classic(base_size = 11) +
  theme(legend.position = "bottom",
        legend.title=element_blank())


```


## plot Fig5A-C

```{r}

ggsave2(paste0("../results/Fig5_A-C.png"),
          plot_grid(p1.1, p3.1, p5.1, 
                    nrow = 1, 
                    ncol = 3, 
                    labels = NULL, 
                    align = "h", 
                    rel_widths = c(1, 1, 1)) + 
          bgcolor("white") + 
          border("white"),
          width = 42, height = 12, unit = "cm")

ggsave2(paste0("../results/Fig5_A-C.pdf"),
          plot_grid(p1.1, p3.1, p5.1, 
                    nrow = 1, 
                    ncol = 3, 
                    labels = NULL, 
                    align = "h", 
                    rel_widths = c(1, 1, 1)) + 
          bgcolor("white") + 
          border("white"),
          width = 42, height = 12, unit = "cm")


```


## ggplot correlations between PCA and variables

```{r}

# test <- glm(pc_out_full[, "PC1"] ~ pc_out_full[, "Age_symptom_onset"])
# test_s <- summary(test)
# test_s$coefficients[2,4]
# 
# test2 <- glm(pc_out_full[, "PC1"] ~ pc_out_full[, "SSA"])

p_pheno2 <- data.frame(row.names = sell)
for (i in c("PC1", "PC2")){
  m <- apply(pc_out_full[, rownames(p_pheno2)], 
             2, 
             function(x){glm(pc_out_full[, i] ~ x)} ) 
  m <- lapply(m, summary)
  p_pheno2[, i] <- c(t(sapply(m, function(x){x$coefficients[2, 4]})))
}

p_pheno2$group <- as.factor(c(rep("var", times = nrow(p_pheno2))))

p_pheno <- rbind(p_pheno1, p_pheno2)

## keep patient group colors
scale_fill_pss <- function(...){
    ggplot2:::manual_scale(
        'color', 
        values = c(CTRL = "#F8766D", 
                   `SSA-` = "#7CAE00",
                   `SSA+` = "#00BFC4",
                   SSAB = "#C77CFF"), 
        ...
    )
}

# function for plotting parameters vs PC1/PC2
plotPCs <- function(df_in, #PCA data frame
                    ctrl_filter, #"yes" if CTRL samples are filtered out
                    pc, #PC1 or PC2
                    pheno_var #names of variables to plot
                    ) {
    p_list <- list()
    for (i in pheno_var){
    
      print(i)
      
      if (pc == "PC1") {
        
        tmp <- select(df_in,
                    i, 
                    PC1,
                    patient_group)
        
      } else if (pc == "PC2") {
        
        tmp <- select(df_in,
                    i, 
                    PC2,
                    patient_group)
        
      }
      
      ifelse(ctrl_filter == "yes", 
             tmp <- tmp[tmp$patient_group != "CTRL", ], 
             tmp <- tmp)

      tmp <- tmp %>% drop_na()
      
      # plot binary variables as boxplots and continous as scatter plots
      ifelse(max(tmp[, 1]) == 1,
      p_list[[i]] <- ggplot(tmp,
                            aes_string(x = as.factor(tmp[, i]),
                                       y = pc)) + 
        geom_boxplot(outlier.shape = NA, na.rm = TRUE) +
        geom_jitter(aes(color = patient_group, na.rm = TRUE)) +
        scale_fill_pss() +
        annotate(geom = "text",
                 x = 0,
                 y = 0.4,
                 hjust = -0.2,
                 label = paste0("P = ", signif(p_pheno[i, pc], digits = 2)), 
                 color = "black") +
        scale_x_discrete(labels = c("0" = "No", 
                                    "1" = "Yes"),
                         drop = FALSE) +
        ggtitle(i) +
        xlab("") +
        ylim(-0.35, 0.45) +
        #stat_compare_means(method = "t.test", label = "p.format", vjust = 0.5, na.rm = TRUE) +
        theme_classic(base_size = 12) +
        theme(plot.title = element_text(size = 11), 
              legend.position = "bottom",
              legend.title = element_blank()), 
      p_list[[i]] <- ggplot(tmp, 
                                aes_string(x = tmp[, i], 
                                           y = pc)) + 
        geom_point(aes(color = patient_group)) +
        scale_fill_pss() +
        ggtitle(i) +
        xlab("") +
        ylim(-0.35, 0.45) +
        geom_smooth(method = glm, se = FALSE, color = "black", weight = 0.5) +
        annotate(geom = "text",
                 x = 0,
                 y = 0.4,
                 hjust = -0.2,
                 label = paste0("P = ", signif(p_pheno[i, pc], digits = 2)), 
                 color = "black") +
        #stat_cor(method = "pearson", na.rm = TRUE) +
        scale_x_continuous(breaks = seq(0, 
                                        signif(max(tmp[, 1]) * 1.1, digits = 1), 
                                        by =  signif((max(tmp[, 1]) * 1.1) / 3, digits = 1)),
                           limits = c(0, signif(max(tmp[, 1]) * 1.1, digits = 1))) +
        theme_classic(base_size = 12) +
        theme(plot.title = element_text(size = 10), 
              legend.position = "bottom",
              legend.title = element_blank())
      )
      
      #p_legend <- get_legend(p_list[[i]]) %>% ggpubr::as_ggplot()
      
      p_list[[i]] <- p_list[[i]] + theme(legend.position = "none")
      
    }
    
    return(p_list)
    
}

p_list_PC1_cpar <- plotPCs(pc_out_full, "yes", "PC1", colnames(pc_out_full)[c(8:10, 13:14, 16:25)])
p_list_PC1_var <- plotPCs(pc_out_full, "no", "PC1", sell)
p_list_PC2_cpar <- plotPCs(pc_out_full, "yes", "PC2", colnames(pc_out_full)[c(8:10, 13:14, 16:25)])
p_list_PC2_var <- plotPCs(pc_out_full, "no", "PC2", sell)

p_legend <- get_legend(p1.1) %>% ggpubr::as_ggplot()


```


## plot SFig19

```{r}

# plot all clinical variables for PC1 and PC2
p1 <- plot_grid(plotlist = p_list_PC1_cpar, 
                    nrow = 5, 
                    ncol = 3, 
                    labels = NULL, 
                    align = "hv") + 
          bgcolor("white") + 
          border("white")

ggsave2(paste0("../results/SFig19.png"),
          plot_grid(p1, p_legend, 
                    ncol = 1, 
                    labels = NULL, 
                    align = "hv",
                    rel_heights = c(10, 0.2)) + 
          bgcolor("white") + 
          border("white"),
          width = 20, height = 30, unit = "cm")

ggsave2(paste0("../results/SFig19.pdf"),
          plot_grid(p1, p_legend, 
                    ncol = 1, 
                    labels = NULL, 
                    align = "hv",
                    rel_heights = c(10, 0.2)) + 
          bgcolor("white") + 
          border("white"),
          width = 20, height = 30, unit = "cm")


```


## plot SFig20

```{r}

p1 <- plot_grid(plotlist = p_list_PC2_cpar, 
                    nrow = 5, 
                    ncol = 3, 
                    labels = NULL, 
                    align = "hv") + 
          bgcolor("white") + 
          border("white")

ggsave2(paste0("../results/SFig20.png"),
          plot_grid(p1, p_legend, 
                    ncol = 1, 
                    labels = NULL, 
                    align = "hv",
                    rel_heights = c(10, 0.2)) + 
          bgcolor("white") + 
          border("white"),
          width = 20, height = 30, unit = "cm")

ggsave2(paste0("../results/SFig20.pdf"),
          plot_grid(p1, p_legend, 
                    ncol = 1, 
                    labels = NULL, 
                    align = "hv",
                    rel_heights = c(10, 0.2)) + 
          bgcolor("white") + 
          border("white"),
          width = 20, height = 30, unit = "cm")


```


## plot SFig21

```{r}

# plot GEX and VDJ variables for PC1 and PC2
#length(pheno_var[16:length(pheno_var)]) #[1] 30
p1 <- plot_grid(plotlist = p_list_PC1_var, 
                    nrow = 6, 
                    ncol = 5, 
                    labels = NULL, 
                    align = "hv") + 
          bgcolor("white") + 
          border("white")

ggsave2(paste0("../results/SFig21.png"),
          plot_grid(p1, p_legend, 
                    ncol = 1, 
                    labels = NULL, 
                    align = "hv",
                    rel_heights = c(10, 0.2)) + 
          bgcolor("white") + 
          border("white"),
          width = 30, height = 30, unit = "cm")

ggsave2(paste0("../results/SFig21.pdf"),
          plot_grid(p1, p_legend, 
                    ncol = 1, 
                    labels = NULL, 
                    align = "hv",
                    rel_heights = c(10, 0.2)) + 
          bgcolor("white") + 
          border("white"),
          width = 30, height = 30, unit = "cm")


```


## plot SFig22

```{r}

p1 <- plot_grid(plotlist = p_list_PC2_var, 
                    nrow = 6, 
                    ncol = 5, 
                    labels = NULL, 
                    align = "hv") + 
          bgcolor("white") + 
          border("white")

ggsave2(paste0("../results/SFig22.png"),
          plot_grid(p1, p_legend, 
                    ncol = 1, 
                    labels = NULL, 
                    align = "hv",
                    rel_heights = c(10, 0.2)) + 
          bgcolor("white") + 
          border("white"),
          width = 30, height = 30, unit = "cm")

ggsave2(paste0("../results/SFig22.pdf"),
          plot_grid(p1, p_legend, 
                    ncol = 1, 
                    labels = NULL, 
                    align = "hv",
                    rel_heights = c(10, 0.2)) + 
          bgcolor("white") + 
          border("white"),
          width = 30, height = 30, unit = "cm")


```


## plot Fig5D-E

```{r}

# plot select patient variables for PC1 
p_pat_var <- c("SSA", "SSB", "ANA", "RF", 
  "Age_symptom_onset", "Age_diagnosis",
  "Focus_score_Greenspan", "P-IgG", "Leucopenia")

p1.1 <- plot_grid(plotlist = p_list_PC1_cpar[p_pat_var], 
                    nrow = 3, 
                    ncol = 3, 
                    labels = NULL, 
                    align = "hv") + 
          bgcolor("white") + 
          border("white")


p1.2 <- plot_grid(p1.1, p_legend, 
                    ncol = 1, 
                    labels = NULL, 
                    align = "hv",
                    rel_heights = c(10, 0.5)) + 
          bgcolor("white") + 
          border("white")


# plot select GEX and VDJ variables for PC1 
p_gex_var <- c("Memory_Classical_IFNscore", "Memory_DN_IFNscore", "Memory_IgM_IFNscore", "Naive_IFNscore",
               "Memory_IgM_VDJmutation", "Transitional_VDJmutation", "IGHM_Naive_%", "IGHM_Transitional_%",
               "IGHV3-69-1_%", "IGHV1-69_%", "Memory_Classical_%", "Memory_IgM_%")

names(p_list_PC1)
p2.1 <- plot_grid(plotlist = p_list_PC1_var[p_gex_var], 
                    nrow = 3, 
                    ncol = 4, 
                    labels = NULL, 
                    align = "hv") + 
          bgcolor("white") + 
          border("white")

p2.2 <- plot_grid(p2.1, p_legend, 
                    ncol = 1, 
                    labels = NULL, 
                    align = "hv",
                    rel_heights = c(10, 0.5)) + 
          bgcolor("white") + 
          border("white")


ggsave2(paste0("../results/Fig5D-E.png"),
          plot_grid(p1.2, NULL, p2.2, 
                    ncol = 3, 
                    labels = NULL, 
                    align = "hv",
                    rel_widths = c(1, 0.1, 1.3)) + 
          bgcolor("white") + 
          border("white"),
          width = 45, height = 18, unit = "cm", dpi = 200)

ggsave2(paste0("../results/Fig5D-E.pdf"),
          plot_grid(p1.2, NULL, p2.2, 
                    ncol = 3, 
                    labels = NULL, 
                    align = "hv",
                    rel_widths = c(1, 0.1, 1.3)) + 
          bgcolor("white") + 
          border("white"),
          width = 45, height = 18, unit = "cm", dpi = 200)


```


## print sessionInfo()

```{r}

sessionInfo()

```
